PowerSystems

PowerSystems

API documentation

PowerSystems

Index

Exported

Internal

Index

PowerSystems.PowerSystems

PowerSystems.conductorless

PowerSystems.Bus

PowerSystems.ConcreteSystem

PowerSystems.DataFormatError

PowerSystems.EconRenewable

PowerSystems.EconThermal

PowerSystems.GeneratorCostModel

PowerSystems.LogEvent

PowerSystems.LogEventTracker

PowerSystems.MultiConductorMatrix

PowerSystems.MultiConductorVector

PowerSystems.MultiLogger

PowerSystems.MultiLogger

PowerSystems.ProportionalReserve

PowerSystems.StaticReserve

PowerSystems.System

PowerSystems.System

PowerSystems.System

PowerSystems.System

PowerSystems.System

PowerSystems.System

PowerSystems.System

PowerSystems.System

PowerSystems.TechThermal

PowerSystems.ThermalDispatch

PowerSystems.ThermalGen

PowerSystems.ThermalGenSeason

PowerSystems.Transformer2W

PowerSystems.VSCDCLine

Base.close

Base.flush

Base.isapprox

Base.setindex!

Base.setindex!

PowerSystems._bold

PowerSystems._calc_max_cost_index

PowerSystems._check_conductors

PowerSystems._check_cost_function

PowerSystems._check_reference_buses

PowerSystems._dfs

PowerSystems._float2string

PowerSystems._get_components_by_type

PowerSystems._grey

PowerSystems._iscomponentdict

PowerSystems._make_mixed_units

PowerSystems._make_multiconductor

PowerSystems._make_per_unit

PowerSystems._propagate_topology_status

PowerSystems._rescale_cost_model

PowerSystems._select_largest_component

PowerSystems._simplify_pwl_cost

PowerSystems._standardize_cost_terms

PowerSystems._summary

PowerSystems._update_data!

PowerSystems._value2string

PowerSystems._value2string

PowerSystems.add_dcline_costs

PowerSystems.add_line_delimiter

PowerSystems.add_section_data!

PowerSystems.add_time_series

PowerSystems.add_time_series_load

PowerSystems.apply_func

PowerSystems.arrays_to_dicts!

PowerSystems.assign_ts_data

PowerSystems.branch_csv_parser

PowerSystems.bus_csv_parser

PowerSystems.bus_gen_lookup

PowerSystems.bus_load_lookup

PowerSystems.bus_shunt_lookup

PowerSystems.calc_branch_t

PowerSystems.calc_branch_y

PowerSystems.calc_max_cost_index

PowerSystems.calc_theta_delta_bounds

PowerSystems.check_ascending_order

PowerSystems.check_branch_directions

PowerSystems.check_branch_loops

PowerSystems.check_bus_types

PowerSystems.check_component_refrence_bus

PowerSystems.check_conductors

PowerSystems.check_connectivity

PowerSystems.check_cost_functions

PowerSystems.check_current_limits

PowerSystems.check_dcline_limits

PowerSystems.check_keys

PowerSystems.check_network_data

PowerSystems.check_reference_buses

PowerSystems.check_storage_parameters

PowerSystems.check_thermal_limits

PowerSystems.check_transformer_parameters

PowerSystems.check_type

PowerSystems.check_voltage_angle_differences

PowerSystems.check_voltage_setpoints

PowerSystems.compare_dict

PowerSystems.compare_numbers

PowerSystems.component_table

PowerSystems.configure_logging

PowerSystems.connected_components

PowerSystems.create_starbus_from_transformer

PowerSystems.csv2ps_dict

PowerSystems.csv2ps_dict

PowerSystems.dc_branch_csv_parser

PowerSystems.export_cost_data

PowerSystems.export_extra_data

PowerSystems.export_matpower

PowerSystems.export_matpower

PowerSystems.extract_matlab_assignment

PowerSystems.find_bus

PowerSystems.find_max_bus_id

PowerSystems.gen_csv_parser

PowerSystems.get_abstract_subtypes

PowerSystems.get_all_concrete_subtypes

PowerSystems.get_bus_value

PowerSystems.get_component_counts

PowerSystems.get_components

PowerSystems.get_concrete_subtypes

PowerSystems.get_default

PowerSystems.get_line_elements

PowerSystems.get_log_events

PowerSystems.get_mixed_components

PowerSystems.get_pti_dtypes

PowerSystems.get_pti_sections

PowerSystems.im_replicate

PowerSystems.import_remaining!

PowerSystems.increment_count

PowerSystems.init_bus!

PowerSystems.ismultinetwork

PowerSystems.load_csv_parser

PowerSystems.load_csv_parser

PowerSystems.loadzone_csv_parser

PowerSystems.make_bus

PowerSystems.make_forecast_array

PowerSystems.make_forecast_dict

PowerSystems.make_forecast_dict

PowerSystems.make_forecast_dict

PowerSystems.make_mixed_units

PowerSystems.make_multiconductor

PowerSystems.make_per_unit

PowerSystems.matpower_to_powermodels

PowerSystems.merge_bus_name_data

PowerSystems.merge_generator_cost_data

PowerSystems.merge_generic_data

PowerSystems.move_genfuel_and_gentype!

PowerSystems.mp2pm_branch

PowerSystems.mp2pm_dcline

PowerSystems.mp_cost_data

PowerSystems.open_file_logger

PowerSystems.parse_file

PowerSystems.parse_json

PowerSystems.parse_json

PowerSystems.parse_line_element!

PowerSystems.parse_matlab_cells

PowerSystems.parse_matlab_data

PowerSystems.parse_matlab_matrix

PowerSystems.parse_matpower

PowerSystems.parse_matpower_file

PowerSystems.parse_matpower_file

PowerSystems.parse_matpower_string

PowerSystems.parse_psse

PowerSystems.parse_psse

PowerSystems.parse_psse

PowerSystems.parse_pti

PowerSystems.parse_pti

PowerSystems.parse_pti_data

PowerSystems.parsestandardfiles

PowerSystems.pm2ps_dict

PowerSystems.print_summary

PowerSystems.print_summary

PowerSystems.propagate_topology_status

PowerSystems.ps_dict2ps_struct

PowerSystems.psse2pm_branch!

PowerSystems.psse2pm_bus!

PowerSystems.psse2pm_dcline!

PowerSystems.psse2pm_generator!

PowerSystems.psse2pm_load!

PowerSystems.psse2pm_shunt!

PowerSystems.psse2pm_transformer!

PowerSystems.read_csv_data

PowerSystems.read_data_files

PowerSystems.read_datetime

PowerSystems.read_gen

PowerSystems.remove_missing

PowerSystems.replicate

PowerSystems.report_log_summary

PowerSystems.report_log_summary

PowerSystems.row_to_dict

PowerSystems.row_to_typed_dict

PowerSystems.select_largest_component

PowerSystems.services_csv_parser

PowerSystems.show_component_counts

PowerSystems.simplify_cost_terms

PowerSystems.split_line

PowerSystems.split_loads_shunts

PowerSystems.standardize_cost_terms

PowerSystems.summary

PowerSystems.summary

PowerSystems.supertypes

PowerSystems.type_array

PowerSystems.type_to_symbol

PowerSystems.type_value

PowerSystems.update_data

PowerSystems.update_data!

PowerSystems.validate

PowerSystems.validate

PowerSystems.validate

PowerSystems.validate

PowerSystems.validate

PowerSystems.validate

PowerSystems.validate

PowerSystems.validate_devices

Exported

PowerSystems.PowerSystems — Module.

Module for constructing self-contained power system objects.

source

PowerSystems.Bus — Type.

Bus

A power-system bus.

Constructor

Bus(number, name, bustype, angle, voltage, voltagelimits, basevoltage)

Arguments

number::Int64 : number associated with the bus

name::String : the name of the bus

bustype::String : type of bus, [PV, PQ, SF]; may be nothing

angle::Float64 : angle of the bus in degrees; may be nothing

voltage::Float64 : voltage as a multiple of basevoltage; may be nothing

voltagelimits::NamedTuple(min::Float64, max::Float64) : limits on the voltage variation as multiples of basevoltage; may be nothing

basevoltage::Float64 : the base voltage in kV; may be nothing

source

PowerSystems.ConcreteSystem — Type.

A System struct that stores all devices in arrays with concrete types. This is a temporary implementation that will allow consumers of PowerSystems to test the functionality before it is finalized.

source

PowerSystems.EconRenewable — Type.

" Data Structure for the economical parameters of renewable generation technologies. The data structure can be called calling all the fields directly or using named fields. All the limits are defined by NamedTuples and some fields can take nothing

## Examples

source

PowerSystems.EconThermal — Type.

" Data Structure for the economical parameters of thermal generation technologies. The data structure can be called calling all the fields directly or using named fields. All the limits are defined by NamedTuples and some fields can take nothing

## Examples

source

PowerSystems.LogEvent — Type.

Contains information describing a log event.

source

PowerSystems.LogEventTracker — Type.

LogEventTracker(Tuple{Logging.LogLevel})

Tracks counts of all log events by level.

Examples

LogEventTracker()
LogEventTracker((Logging.Info, Logging.Warn, Logging.Error))

source

PowerSystems.MultiConductorMatrix — Type.

source

PowerSystems.MultiConductorVector — Type.

a data structure for working with multiconductor datasets

source

PowerSystems.MultiLogger — Type.

MultiLogger(Array{AbstractLogger}, Union{LogEventTracker, Nothing})

Redirects log events to multiple loggers. The primary use case is to allow logging to both a file and the console. Secondarily, it can track the counts of all log messages.

Example

MultiLogger([ConsoleLogger(stderr), SimpleLogger(stream)], LogEventTracker())

source

PowerSystems.MultiLogger — Method.

Creates a MultiLogger with no event tracking.

Example

MultiLogger([ConsoleLogger(stderr), SimpleLogger(stream)])

source

PowerSystems.ProportionalReserve — Type.

ProportionalReserve(name::String, contributingdevices::Device, timeframe::Float64, requirement::Dict{Any,Dict{Int,TimeSeries.TimeArray}})

Data Structure for a proportional reserve product for system simulations. The data structure can be called calling all the fields directly or using named fields. name - description contributingdevices - devices from which the product can be procured timeframe - the relative saturation timeframe requirement - the required quantity of the product

Examples

```jldoctest

source

PowerSystems.StaticReserve — Type.

StaticReserve(name::String, contributingdevices::Device, timeframe::Float64, requirement::Float64})

Data Structure for the procurement products for system simulations. The data structure can be called calling all the fields directly or using named fields. name - description contributingdevices - devices from which the product can be procured timeframe - the relative saturation timeframe requirement - the required quantity of the product

Examples

```jldoctest

source

PowerSystems.System — Type.

System

A power system defined by fields for buses, generators, loads, branches, and overall system parameters.

Constructor

System(buses, generators, loads, branches, storage, basepower; kwargs...)
System(buses, generators, loads, branches, basepower; kwargs...)
System(buses, generators, loads, basepower; kwargs...)
System(ps_dict; kwargs...)
System(file, ts_folder; kwargs...)
System(; kwargs...)

Arguments

buses::Vector{Bus} : an array of buses

generators::Vector{Generator} : an array of generators of (possibly) different types

loads::Vector{ElectricLoad} : an array of load specifications that includes timing of the loads

branches::Union{Nothing, Vector{Branch}} : an array of branches; may be nothing

storage::Union{Nothing, Vector{Storage}} : an array of storage devices; may be nothing

basepower::Float64 : the base power of the system (DOCTODO: is this true? what are the units of base power?)

ps_dict::Dict{String,Any} : the dictionary object containing System data

file::String, ts_folder::String : the filename and foldername that contain the System data

Keyword arguments

runchecks::Bool : run available checks on input fields

DOCTODO: any other keyword arguments? genmapfile, REGEXFILE

source

PowerSystems.System — Method.

Primary System constructor. Funnel point for all other outer constructors.

source

PowerSystems.System — Method.

Constructs System with Generators but no storage.

source

PowerSystems.System — Method.

Constructs System with Generators but no branches.

source

PowerSystems.System — Method.

Constructs System with Generators but no branches or storage.

source

PowerSystems.System — Method.

Constructs System from a ps_dict.

source

PowerSystems.System — Method.

Constructs System from a file containing Matpower, PTI, or JSON data.

source

PowerSystems.System — Method.

Constructs System with default values.

source

PowerSystems.TechThermal — Type.

TechThermal(activepower::Float64,
        activepowerlimits::NamedTuple{(:min, :max),Tuple{Float64,Float64}},
        reactivepower::Union{Float64,Nothing},
        reactivepowerlimits::Union{(min::Float64,max::Float64),Nothing},
        ramplimits::Union{NamedTuple{(:up, :down),Tuple{Float64,Float64}},Nothing},
        timelimits::Union{NamedTuple{(:min, :max),Tuple{Float64,Float64}},Nothing})

Data Structure for the economical parameters of thermal generation technologies. The data structure can be called calling all the fields directly or using named fields. Two examples [DOCTODO only one example exists below] are provided one with minimal data definition and a more comprenhensive one

# Examples

```jldoctest

julia> Tech = TechThermal(activepower = 100.0, activepowerlimits = (min = 50.0, max = 200.0))

[ Info: 'Reactive Power' limits defined as nothing TechThermal: activepower: 100.0 activepowerlimits: (min = 50.0, max = 200.0) reactivepower: nothing reactivepowerlimits: nothing ramplimits: nothing timelimits: nothing

source

PowerSystems.ThermalDispatch — Type.

" Data Structure for thermal generation technologies. The data structure contains all the information for technical and economical modeling. The data fields can be filled using named fields or directly.

Examples

source

PowerSystems.ThermalGen — Type.

Abstract struct for thermal generation technologies

source

PowerSystems.ThermalGenSeason — Type.

" Data Structure for thermal generation technologies subjecto to seasonality constraints. The data structure contains all the information for technical and economical modeling and an extra field for a time series. The data fields can be filled using named fields or directly.

Examples

source

PowerSystems.Transformer2W — Type.

The 2-W transformer model uses an equivalent circuit assuming the impedance is on the High Voltage Side of the transformer. The model allocates the iron losses and magnetezing suceptance to the primary side

source

PowerSystems.VSCDCLine — Type.

As implemented in Milano's Book Page 397

source

PowerSystems.assign_ts_data — Method.

Args: PowerSystems Dictionary Dictionary of all the data files Returns: Returns an dictionary with Device name as key and PowerSystems Forecasts dictionary as values

source

PowerSystems.configure_logging — Method.

configure_logging([console, console_stream, console_level,
                   file, filename, file_level, file_mode,
                   tracker, set_global])

Creates console and file loggers per caller specification and returns a MultiLogger.

Note: If logging to a file users must call Base.close() on the returned MultiLogger to ensure that all events get flushed.

Arguments

console::Bool=true: create console logger

console_stream::IOStream=stderr: stream for console logger

console_level::Logging.LogLevel=Logging.Error: level for console messages

file::Bool=true: create file logger

filename::String=log.txt: log file

file_level::Logging.LogLevel=Logging.Info: level for file messages

file_mode::String=w+: mode used when opening log file

tracker::Union{LogEventTracker, Nothing}=LogEventTracker(): optionally track log events

set_global::Bool=true: set the created logger as the global logger

Example

logger = configure_logging(filename="mylog.txt")

source

PowerSystems.get_component_counts — Method.

Returns a Tuple of Arrays of component types and counts.

source

PowerSystems.get_components — Method.

Returns an array of components from the System. T must be a concrete type.

Example

devices = PowerSystems.get_components(ThermalDispatch, system)

source

PowerSystems.get_log_events — Method.

Returns an iterable of log events for a level.

source

PowerSystems.get_mixed_components — Method.

Returns an iterable over component arrays that are subtypes of T. To create a new array with all component arrays concatenated, call collect on the returned iterable. Note that this will involve copying the data and the resulting array will be less performant than arrays of concrete types.

Example

iter = PowerSystems.get_mixed_components(Device, system)
for device in iter
    @show device
end

source

PowerSystems.open_file_logger — Function.

open_file_logger(func, filename[, level, mode])

Opens a file logger using Logging.SimpleLogger.

Example

open_file_logger("log.txt", Logging.Info) do logger
    global_logger(logger)
    @info "hello world"
end

source

PowerSystems.parse_file — Method.

parse_file(file; import_all)

Parses a Matpower .m file or PTI (PSS(R)E-v33) .raw file into a PowerModels data structure. All fields from PTI files will be imported if import_all is true (Default: false).

source

PowerSystems.parsestandardfiles — Method.

Read in power-system parameters from a Matpower, PTI, or JSON file and do some data checks.

source

PowerSystems.ps_dict2ps_struct — Method.

Takes a PowerSystems dictionary and return an array of PowerSystems struct for Bus, Generator, Branch and load

source

PowerSystems.read_data_files — Method.

Read all forecast CSV's in the path provided, the struct of the data should follow this format folder : PV file : DAYAHEAD file : REALTIME Folder name should be the device type Files should only contain one real-time and day-ahead forecast Args: files: A string Returns: A dictionary with the CSV files as dataframes and folder names as keys

TODO : Stochasti/Multiple scenarios

source

PowerSystems.report_log_summary — Method.

Returns a summary of log event counts by level.

source

PowerSystems.report_log_summary — Method.

Returns a summary of log event counts by level.

source

PowerSystems.show_component_counts — Function.

Shows the component types and counts in a table. If show_hierarchy is true then include a column showing the type hierachy. The display is in order of depth-first type hierarchy.

source

PowerSystems.update_data! — Method.

recursively applies new_data to data, overwriting information

source

PowerSystems.validate — Method.

Validates the contents of a Branch.

source

PowerSystems.validate — Method.

Validates the contents of a Bus.

source

PowerSystems.validate — Method.

Validates the contents of an ElectricLoad.

source

PowerSystems.validate — Method.

Validates the contents of generators.

source

PowerSystems.validate — Method.

Validates the contents of a Generator.

source

PowerSystems.validate — Method.

Validates the contents of a Storage.

source

PowerSystems.validate — Method.

Validates the contents of a System.

source

Internal

PowerSystems.conductorless — Constant.

feild names that should not be multi-conductor values

source

PowerSystems.DataFormatError — Type.

Thrown upon detection of user data that is not supported.

source

PowerSystems.GeneratorCostModel — Type.

From http://www.pserc.cornell.edu/matpower/MATPOWER-manual.pdf Table B-4

source

Base.close — Method.

Ensures that any file streams are flushed and closed.

source

Base.flush — Method.

Flush any file streams.

source

Base.isapprox — Method.

source

Base.setindex! — Method.

source

Base.setindex! — Method.

source

PowerSystems._bold — Method.

Makes a string bold in the terminal

source

PowerSystems._calc_max_cost_index — Method.

source

PowerSystems._check_conductors — Method.

source

PowerSystems._check_cost_function — Method.

source

PowerSystems._check_reference_buses — Method.

source

PowerSystems._dfs — Method.

performs DFS on a graph

source

PowerSystems._float2string — Method.

converts a float value into a string of fixed precision

sprintf would do the job but this work around is needed because sprintf cannot take format strings during runtime

source

PowerSystems._get_components_by_type — Method.

Builds a nested dictionary by traversing through the PowerSystems type hierarchy. The bottom of each dictionary is an array of concrete types.

Example

data[Device][Injection][Generator][ThermalGen][ThermalDispatch][1]
ThermalDispatch:
   name: 322_CT_6
   available: true
   bus: Bus(name="Cole")
   tech: TechThermal
   econ: EconThermal

source

PowerSystems._grey — Method.

Makes a string grey in the terminal, does not seem to work well on Windows terminals more info can be found at https://en.wikipedia.org/wiki/ANSIescapecode

source

PowerSystems._iscomponentdict — Method.

Attempts to determine if the given data is a component dictionary

source

PowerSystems._make_mixed_units — Method.

source

PowerSystems._make_multiconductor — Method.

source

PowerSystems._make_per_unit — Method.

source

PowerSystems._propagate_topology_status — Method.

source

PowerSystems._rescale_cost_model — Method.

source

PowerSystems._select_largest_component — Method.

source

PowerSystems._simplify_pwl_cost — Function.

checks the slope of each segment in a pwl function, simplifies the function if the slope changes is below a tolerance

source

PowerSystems._standardize_cost_terms — Method.

ensures all polynomial costs functions have at exactly comp_order terms

source

PowerSystems._summary — Method.

prints the text summary for a data dictionary to IO

source

PowerSystems._update_data! — Method.

recursive call of updatedata

source

PowerSystems._value2string — Method.

converts any value to a string, summarizes arrays and dicts

source

PowerSystems._value2string — Method.

converts a MultiConductorValue value to a string in summary

source

PowerSystems.add_dcline_costs — Method.

adds dcline costs, if gen costs exist

source

PowerSystems.add_line_delimiter — Method.

source

PowerSystems.add_section_data! — Method.

add_section_data!(pti_data, section_data, section)

Adds section_data::Dict, which contains all parsed elements of a PTI file section given by section, into the parent pti_data::Dict

source

PowerSystems.add_time_series — Method.

Arg: Device dictionary - Generators Dataframe contains device Realtime/Forecast TimeSeries Returns: Device dictionary with timeseries added

source

PowerSystems.add_time_series_load — Method.

Arg: Load dictionary LoadZones dictionary Dataframe contains device Realtime/Forecast TimeSeries Returns: Device dictionary with timeseries added

source

PowerSystems.apply_func — Method.

source

PowerSystems.arrays_to_dicts! — Method.

turns top level arrays into dicts

source

PowerSystems.branch_csv_parser — Function.

Args: A DataFrame with the same column names as in RTS_GMLC branch.csv file Parsed Bus PowerSystems dictionary Returns: A Nested Dictionary with keys as branch types/names and values as line/transformer data dictionary with same keys as the device struct

source

PowerSystems.bus_csv_parser — Function.

Args: A DataFrame with the same column names as in RTS_GMLC bus.csv file

"Bus ID" "Bus Name" "BaseKV" "Bus Type" "MW Load" "MVAR Load" "V Mag" "V
Angle" "MW Shunt G" "MVAR Shunt B" "Area"

Returns: A Nested Dictionary with keys as Bus number and values as bus data dictionary with same keys as the device struct

source

PowerSystems.bus_gen_lookup — Method.

builds a lookup list of what generators are connected to a given bus

source

PowerSystems.bus_load_lookup — Method.

builds a lookup list of what loads are connected to a given bus

source

PowerSystems.bus_shunt_lookup — Method.

builds a lookup list of what shunts are connected to a given bus

source

PowerSystems.calc_branch_t — Method.

source

PowerSystems.calc_branch_y — Method.

source

PowerSystems.calc_max_cost_index — Method.

source

PowerSystems.calc_theta_delta_bounds — Method.

source

PowerSystems.check_ascending_order — Method.

Throws DataFormatError if the array is not in ascending order.

source

PowerSystems.check_branch_directions — Method.

checks that all parallel branches have the same orientation

source

PowerSystems.check_branch_loops — Method.

checks that all branches connect two distinct buses

source

PowerSystems.check_bus_types — Method.

checks bus types are consistent with generator connections, if not, fixes them

source

PowerSystems.check_component_refrence_bus — Method.

checks that a connected component has a reference bus, if not, adds one

source

PowerSystems.check_conductors — Method.

source

PowerSystems.check_connectivity — Method.

checks that all buses are unique and other components link to valid buses

source

PowerSystems.check_cost_functions — Method.

throws warnings if cost functions are malformed

source

PowerSystems.check_current_limits — Method.

checks that each branch has a reasonable current rating-a, if not computes one

source

PowerSystems.check_dcline_limits — Method.

checks that parameters for dc lines are reasonable

source

PowerSystems.check_keys — Method.

source

PowerSystems.check_network_data — Method.

Runs various data quality checks on a PowerModels data dictionary. Applies modifications in some cases. Reports modified component ids.

source

PowerSystems.check_reference_buses — Method.

checks that each connected components has a reference bus, if not, adds one

source

PowerSystems.check_storage_parameters — Method.

checks that each storage unit has a reasonable parameters

source

PowerSystems.check_thermal_limits — Method.

checks that each branch has a reasonable thermal rating-a, if not computes one

source

PowerSystems.check_transformer_parameters — Method.

checks that each branch has a reasonable transformer parameters

this is important because setting tap == 0.0 leads to NaN computations, which are hard to debug

source

PowerSystems.check_type — Method.

Checks if the given value is of a given type, if not tries to make it that type

source

PowerSystems.check_voltage_angle_differences — Function.

checks that voltage angle differences are within 90 deg., if not tightens

source

PowerSystems.check_voltage_setpoints — Method.

throws warnings if generator and dc line voltage setpoints are not consistent with the bus voltage setpoint

source

PowerSystems.compare_dict — Method.

tests if two dicts are equal, up to floating point precision

source

PowerSystems.compare_numbers — Method.

tests if two numbers are equal, up to floating point precision

source

PowerSystems.component_table — Method.

builds a table of component data

source

PowerSystems.connected_components — Method.

computes the connected components of the network graph returns a set of sets of bus ids, each set is a connected component

source

PowerSystems.create_starbus_from_transformer — Method.

create_starbus(pm_data, transformer)

Creates a starbus from a given three-winding transformer. "sourceid" is given by `["busi", "name", "I", "J", "K", "CKT"]` where "bus_i" and "name" are the modified names for the starbus, and "I", "J", "K" and "CKT" come from the originating transformer, in the PSS(R)E transformer specification.

source

PowerSystems.csv2ps_dict — Method.

Args: Dict with all the System data from CSV files ... see read_csv_data()' Returns: A Power Systems Nested dictionary with keys as devices and values as data dictionary necessary to construct the device structs PS dictionary: "Bus" => Dict(bus_no => Dict("name" => "number" => ... ) ) "Generator" => Dict( "Thermal" => Dict( "name" => "tech" => ...) "Hydro" => .. "Renewable" => .. ) "Branch" => ... "Load" => ... "LoadZones" => ... "BaseKV" => .. ...

source

PowerSystems.csv2ps_dict — Method.

Args: Path to folder with all the System data CSV's files Returns: A Power Systems Nested dictionary with keys as devices and values as data dictionary necessary to construct the device structs PS dictionary: "Bus" => Dict(bus_no => Dict("name" => "number" => ... ) ) "Generator" => Dict( "Thermal" => Dict( "name" => "tech" => ...) "Hydro" => .. "Renewable" => .. ) "Branch" => ... "Load" => ... "LoadZones" => ... "BaseKV" => .. ...

source

PowerSystems.dc_branch_csv_parser — Function.

Args: A DataFrame with the same column names as in RTSGMLC dcbranch.csv file Parsed Bus PowerSystems dictionary Returns: A Nested Dictionary with keys as dcbranch types/names and values as dcbranch data dictionary with same keys as the device struct

source

PowerSystems.export_cost_data — Method.

Export cost data

source

PowerSystems.export_extra_data — Function.

Export fields of a component type

source

PowerSystems.export_matpower — Method.

Export power network data in the matpower format

source

PowerSystems.export_matpower — Method.

Export power network data in the matpower format

source

PowerSystems.extract_matlab_assignment — Method.

breaks up matlab strings of the form 'name = value;'

source

PowerSystems.find_bus — Method.

Finds the bus dictionary where a Generator/Load is located or the from & to bus for a line/transformer

source

PowerSystems.find_max_bus_id — Method.

find_max_bus_id(pm_data)

Returns the maximum bus id in pm_data

source

PowerSystems.gen_csv_parser — Function.

Args: A DataFrame with the same column names as in RTS_GMLC gen.csv file Parsed Bus PowerSystems dictionary Returns: A Nested Dictionary with keys as generator types/names and values as generator data dictionary with same keys as the device struct

source

PowerSystems.get_abstract_subtypes — Method.

Returns an array of abstract types that are direct subtypes of T.

source

PowerSystems.get_all_concrete_subtypes — Method.

Returns an array of all concrete subtypes of T.

source

PowerSystems.get_bus_value — Method.

get_bus_value(bus_i, field, pm_data)

Returns the value of field of bus_i from the PowerModels data. Requires "bus" Dict to already be populated.

source

PowerSystems.get_concrete_subtypes — Method.

Returns an array of concrete types that are direct subtypes of T.

source

PowerSystems.get_default — Function.

Get a default value for dict entry

source

PowerSystems.get_line_elements — Method.

get_line_elements(line)

Uses regular expressions to extract all separate data elements from a line of a PTI file and populate them into an Array{String}. Comments, typically indicated at the end of a line with a '/' character, are also extracted separately, and Array{Array{String}, String} is returned.

source

PowerSystems.get_pti_dtypes — Method.

get_pti_dtypes(field_name)

Returns OrderedDict of data types for PTI file section given by field_name, as enumerated by PSS/E Program Operation Manual

source

PowerSystems.get_pti_sections — Method.

get_pti_sections()

Returns Array of the names of the sections, in the order that they appear in a PTI file, v33+

source

PowerSystems.im_replicate — Method.

Transforms a single network into a multinetwork with several deepcopies of the original network

source

PowerSystems.import_remaining! — Method.

Imports remaining keys from data_in into data_out, excluding keys in exclude

source

PowerSystems.increment_count — Method.

Increments the count of a log event.

source

PowerSystems.init_bus! — Method.

init_bus!(bus, id)

Initializes a bus of id id with default values given in the PSS(R)E specification.

source

PowerSystems.ismultinetwork — Method.

checks if a given network data is a multinetwork

source

PowerSystems.load_csv_parser — Function.

Args: A DataFrame with the same column names as in RTS_GMLC bus.csv file Parsed Bus entry of PowerSystems dictionary Optional Args: DataFrame of LoadZone timeseries data Dict of bus column names Dict of load LoadZone timeseries column names Returns: A Nested Dictionary with keys as load names and values as load data dictionary with same keys as the device struct

source

PowerSystems.load_csv_parser — Function.

Args: A DataFrame with the same column names as in RTS_GMLC bus.csv file Parsed Bus entry of PowerSystems dictionary Parsed LoadZone entry of PowerSystems dictionary Optional Args: DataFrame of LoadZone timeseries data Dict of bus column names Dict of load LoadZone timeseries column names Returns: A Nested Dictionary with keys as load names and values as load data dictionary with same keys as the device struct

source

PowerSystems.loadzone_csv_parser — Function.

Args: A DataFrame with the same column names as in RTS_GMLC bus.csv file Parsed Bus PowerSystems dictionary Returns: A Nested Dictionary with keys as loadzone names and values as loadzone data dictionary with same keys as the device struct

source

PowerSystems.make_bus — Method.

Creates a PowerSystems.Bus from a PowerSystems bus dictionary

source

PowerSystems.make_forecast_array — Method.

Args: A PowerSystems forecast dictionary Returns: A PowerSystems forecast stuct array

source

PowerSystems.make_forecast_dict — Method.

Args: Dictionary of all the data files Length of the forecast - Week()/Dates.Day()/Dates.Hour() Forecast horizon in hours - Int64 Array of PowerSystems devices in the systems- Loads Array of PowerSystems LoadZones

Returns: Returns an dictionary with Device name as key and PowerSystems Forecasts dictionary as values

source

PowerSystems.make_forecast_dict — Method.

Args: Dictionary of all the data files Length of the forecast - Week()/Dates.Day()/Dates.Hour() Forecast horizon in hours - Int64 Array of PowerSystems devices in the systems- Loads Returns: Returns an dictionary with Device name as key and PowerSystems Forecasts dictionary as values

source

PowerSystems.make_forecast_dict — Method.

Args: Dictionary of all the data files Length of the forecast - Week()/Dates.Day()/Dates.Hour() Forecast horizon in hours - Int64 Array of PowerSystems devices in the systems - Renewable Generators and Loads Returns: Returns an dictionary with Device name as key and PowerSystems Forecasts dictionary as values

source

PowerSystems.make_mixed_units — Method.

Transforms network data into mixed-units (inverse of per-unit)

source

PowerSystems.make_multiconductor — Method.

Transforms single-conductor network data into multi-conductor data

source

PowerSystems.make_per_unit — Method.

Transforms network data into per-unit

source

PowerSystems.matpower_to_powermodels — Method.

Converts a Matpower dict into a PowerModels dict

source

PowerSystems.merge_bus_name_data — Method.

merges bus name data into buses, if names exist

source

PowerSystems.merge_generator_cost_data — Method.

merges generator cost functions into generator data, if costs exist

source

PowerSystems.merge_generic_data — Method.

merges Matpower tables based on the table extension syntax

source

PowerSystems.move_genfuel_and_gentype! — Method.

Move gentype and genfuel fields to be subfields of gen

source

PowerSystems.mp2pm_branch — Method.

sets all branch transformer taps to 1.0, to simplify branch models

source

PowerSystems.mp2pm_dcline — Method.

adds pmin and pmax values at to and from buses

source

PowerSystems.mp_cost_data — Method.

source

PowerSystems.parse_json — Method.

source

PowerSystems.parse_json — Method.

source

PowerSystems.parse_line_element! — Method.

parse_line_element!(data, elements, section)

Parses a single "line" of data elements from a PTI file, as given by elements which is an array of the line, typically split at ,. Elements are parsed into data types given by section and saved into data::Dict

source

PowerSystems.parse_matlab_cells — Method.

source

PowerSystems.parse_matlab_data — Method.

source

PowerSystems.parse_matlab_matrix — Method.

source

PowerSystems.parse_matpower — Method.

Parses the matpwer data from either a filename or an IO object

source

PowerSystems.parse_matpower_file — Method.

source

PowerSystems.parse_matpower_file — Method.

source

PowerSystems.parse_matpower_string — Method.

source

PowerSystems.parse_psse — Method.

parse_psse(pti_data)

Converts PSS(R)E-style data parsed from a PTI raw file, passed by pti_data into a format suitable for use internally in PowerModels. Imports all remaining data from the PTI file if import_all is true (Default: false).

source

PowerSystems.parse_psse — Method.

Parses directly from iostream

source

PowerSystems.parse_psse — Method.

Parses directly from file

source

PowerSystems.parse_pti — Method.

parse_pti(io::IO)

Reads PTI data in io::IO, returning a Dict of the data parsed into the proper types.

source

PowerSystems.parse_pti — Method.

parse_pti(filename::String)

Open PTI raw file given by filename, returning a Dict of the data parsed into the proper types.

source

PowerSystems.parse_pti_data — Method.

parse_pti_data(data_string, sections)

Parse a PTI raw file into a Dict, given the data_string of the file and a list of the sections in the PTI file (typically given by default by get_pti_sections().

source

PowerSystems.pm2ps_dict — Method.

Takes a dictionary parsed by PowerModels and returns a PowerSystems dictionary. Currently Supports MATPOWER and PSSE data files parsed by PowerModels

source

PowerSystems.print_summary — Method.

prints the text summary for a data dictionary to stdout

source

PowerSystems.print_summary — Method.

prints the text summary for a data file or dictionary to stdout

source

PowerSystems.propagate_topology_status — Method.

finds active network buses and branches that are not necessary for the computation and sets their status to off.

Works on a PowerModels data dict, so that a it can be used without a GenericPowerModel object

Warning: this implementation has quadratic complexity, in the worst case

source

PowerSystems.psse2pm_branch! — Method.

psse2pm_branch!(pm_data, pti_data)

Parses PSS(R)E-style Branch data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "CKT"] in PSS(R)E Branch specification.

source

PowerSystems.psse2pm_bus! — Method.

psse2pm_bus!(pm_data, pti_data)

Parses PSS(R)E-style Bus data into a PowerModels-style Dict. "source_id" is given by ["I", "NAME"] in PSS(R)E Bus specification.

source

PowerSystems.psse2pm_dcline! — Method.

psse2pm_dcline!(pm_data, pti_data)

Parses PSS(R)E-style Two-Terminal and VSC DC Lines data into a PowerModels compatible Dict structure by first converting them to a simple DC Line Model. For Two-Terminal DC lines, "sourceid" is given by ["IPR", "IPI", "NAME"] in the PSS(R)E Two-Terminal DC specification. For Voltage Source Converters, "sourceid" is given by ["IBUS1", "IBUS2", "NAME"], where "IBUS1" is "IBUS" of the first converter bus, and "IBUS2" is the "IBUS" of the second converter bus, in the PSS(R)E Voltage Source Converter specification.

source

PowerSystems.psse2pm_generator! — Method.

psse2pm_generator!(pm_data, pti_data)

Parses PSS(R)E-style Generator data in a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in PSS(R)E Generator specification.

source

PowerSystems.psse2pm_load! — Method.

psse2pm_load!(pm_data, pti_data)

Parses PSS(R)E-style Load data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] in the PSS(R)E Load specification.

source

PowerSystems.psse2pm_shunt! — Method.

psse2pm_shunt!(pm_data, pti_data)

Parses PSS(R)E-style Fixed and Switched Shunt data into a PowerModels-style Dict. "source_id" is given by ["I", "ID"] for Fixed Shunts, and ["I", "SWREM"] for Switched Shunts, as given by the PSS(R)E Fixed and Switched Shunts specifications.

source

PowerSystems.psse2pm_transformer! — Method.

psse2pm_transformer!(pm_data, pti_data)

Parses PSS(R)E-style Transformer data into a PowerModels-style Dict. "source_id" is given by ["I", "J", "K", "CKT", "winding"], where "winding" is 0 if transformer is two-winding, and 1, 2, or 3 for three-winding, and the remaining keys are defined in the PSS(R)E Transformer specification.

source

PowerSystems.read_csv_data — Method.

Reads in all the data stored in csv files The general format for data is folder: gen.csv branch.csv bus.csv .. load.csv Args: Path to folder with all the System data CSV files

Returns: Nested Data dictionary with key values as folder/file names and dataframes as values

source

PowerSystems.read_datetime — Method.

Arg: Dataframes which includes a timerseries columns of either: Year, Month, Day, Period or DateTime or nothing (creates a today referenced DateTime Column) Returns: Dataframe with a DateTime columns

source

PowerSystems.read_gen — Method.

Transfer generators to ps_dict according to their classification

source

PowerSystems.remove_missing — Method.

Arg: Any DataFrame with Missing values / "NA" strings that are either created by readtable() or CSV.read() Returns: DataFrame with missing values replaced by 0

source

PowerSystems.replicate — Method.

calls the replicate function with PowerModels' global keys

source

PowerSystems.row_to_dict — Method.

takes a row from a matrix and assigns the values names

source

PowerSystems.row_to_typed_dict — Method.

takes a row from a matrix and assigns the values names and types

source

PowerSystems.select_largest_component — Method.

determines the largest connected component of the network and turns everything else off

source

PowerSystems.services_csv_parser — Function.

Args: A DataFrame with the same column names as in RTSGMLC reserves.csv file A DataFrame with the same column names as in RTSGMLC gen.csv file A DataFrame with the same column names as in RTS_GMLC bus.csv file Returns: A Nested Dictionary with keys as loadzone names and values as loadzone data dictionary with same keys as the device struct

source

PowerSystems.simplify_cost_terms — Method.

trims zeros from higher order cost terms

source

PowerSystems.split_line — Method.

source

PowerSystems.split_loads_shunts — Method.

split_loads_shunts(data)

Seperates Loads and Shunts in data under separate "load" and "shunt" keys in the PowerModels data format. Includes references to originating bus via "loadbus" and "shuntbus" keys, respectively.

source

PowerSystems.standardize_cost_terms — Method.

ensures all polynomial costs functions have the same number of terms

source

PowerSystems.summary — Method.

prints the text summary for a data dictionary to IO

source

PowerSystems.summary — Method.

prints the text summary for a data file to IO

source

PowerSystems.supertypes — Method.

Returns an array of all super types of T.

source

PowerSystems.type_array — Method.

Attempts to determine the type of an array of strings extracted from a matlab file

source

PowerSystems.type_to_symbol — Method.

Converts a DataType to a Symbol, stripping off the module name(s).

source

PowerSystems.type_value — Method.

Attempts to determine the type of a string extracted from a matlab file

source

PowerSystems.update_data — Method.

recursively applies new_data to data, overwriting information

source

PowerSystems.validate_devices — Method.

Validates an array of devices.

source