Internal API

HydroPowerSimulations.ActivePowerPumpOutageConstraint — Type

Constraint to limit the active power pump variable during an event

HydroPowerSimulations.EnergyBalanceExpression — Type

Expression for PowerSystems.System that keep track of the energy balance for the system in medium term planning

HydroPowerSimulations.EnergyCapacityTimeSeriesParameter — Type

Parameter to define energy capacity limits for hydro pump-turbine time series

HydroPowerSimulations.FeedForwardHydroUsageLimitConstraint — Type

Struct to create the constraint that limits the hydro usage for hydro formulations.

For more information check HydroPowerSimulations Formulations.

The specified constraint is formulated as:

\[\sum_{t=1}^T E^\text{hy}_t \le \text{HydroUsageLimitParameter}_T,\]

HydroPowerSimulations.HydroPowerConstraint — Type

Struct to model turbine outflow limits

For more information check HydroPowerSimulations Formulations.

The specified constraint is formulated as:

\[\ p_{t} = \Delta t (f^{Tu}_{t-1}(0.5 K_1 (v_{t} + v_{t-1}) + K_2))\]

HydroPowerSimulations.LevelTargetFeedforward — Type

LevelTargetFeedforward(
    component_type::Type{<:PowerSystems.Component},
    source::Type{T},
    affected_values::Vector{DataType},
    meta = CONTAINER_KEY_EMPTY_META
) where {T}

Adds a constraint to enforce a minimum level (head or volume) for reservoirs based on the source value. It is recommended to use the Volume or Head variable as Source, affecting the LevelTargetParameter.

Arguments:

component_type::Type{<:PowerSystems.Component} : Specify the type of component on which the Feedforward will be applied

source::Type{T} : Specify the VariableType, ParameterType or AuxVariableType as the source of values for the Feedforward

affected_values::Vector{DataType} : Specify the variable on which the hydro limit will be applied using the source values

HydroPowerSimulations.LevelTargetParameter — Type

Parameter to define the level target for feedforward

HydroPowerSimulations.TurbineFlowLimitConstraint — Type

Struct to limit the turbine flow

\[QW^{min} \le \sum_{j \in J(i)}^T wq_{jt} \le QW^{max},\]

PowerSimulations.add_constraints! — Method

This function defines the constraints for the energy balance in a medium term planning problem.

PowerSimulations.add_constraints! — Method

This function defines the constraint for the hydro power generation for the HydroWaterFactorModel.

PowerSimulations.add_constraints! — Method

This function define the budget constraint for the active power budget formulation. $sum(P[t]) <= Budget$

PowerSimulations.add_constraints! — Method

This function defines the constraints for the energy level for the PowerSystems.HydroReservoir using the HydroEnergyModelReservoir formulation.

PowerSimulations.add_constraints! — Method

This function defines the constraints for the water level (or state of charge) for the HydroWaterFactorModel.

PowerSimulations.add_constraints! — Method

This function define the relationship between turbined flow and power produced

PowerSimulations.add_constraints! — Method

This function define the relationship between turbined flow and power produced with constant head

PowerSimulations.add_constraints! — Method

This function define the budget constraint for the active power budget formulation. $sum(f[t]) <= Budget$

PowerSimulations.add_constraints! — Method

Add power variable limits constraints for abstract hydro dispatch formulations

PowerSimulations.add_constraints! — Method

Add power variable limits constraints for abstract hydro unit commitment formulations

PowerSimulations.add_constraints! — Method

Add reactive power variable limits constraints for hydro pump formulations

PowerSimulations.add_constraints! — Method

This function defines the constraints for the pump power for the PowerSystems.HydroPumpTurbine.

PowerSimulations.add_constraints! — Method

This function define the budget constraint for the active power budget formulation. $sum(P[t]) <= Budget$

PowerSimulations.add_constraints! — Method

Add energy target constraints for PowerSystems.HydroGen

PowerSimulations.add_constraints! — Method

Add energy target constraints for PowerSystems.HydroReservoir

PowerSimulations.add_constraints! — Method

Add semicontinuous LB range constraints for HydroPumpEnergyCommitment formulation

PowerSimulations.add_constraints! — Method

Add semicontinuous LB range constraints for HydroPumpEnergyDispatch formulation

PowerSimulations.add_constraints! — Method

Add semicontinuous UB range constraints for HydroPumpEnergyCommitment formulation

PowerSimulations.add_constraints! — Method

Add semicontinuous UB range constraints for HydroPumpEnergyDispatch formulation

PowerSimulations.add_constraints! — Method

Time series constraints

PowerSimulations.add_constraints! — Method

Add semicontinuous range constraints for HydroTurbineWaterLinearCommitment formulation

PowerSimulations.add_constraints! — Method

Add semicontinuous range constraints for HydroCommitmentRunOfRiver formulation

PowerSimulations.add_constraints! — Method

This function define the relationship between head and volume for the reservoir.

PowerSimulations.add_constraints! — Method

This function define the level (head or volume) limits for the reservoir.

PowerSimulations.add_constraints! — Method

This function define the level (head or volume) limits for the reservoir.

PowerSimulations.add_constraints! — Method

This function define the target level constraint for the reservoir.

PowerSimulations.add_constraints! — Method

This function define the target level constraint for the reservoir.

PowerSimulations.add_constraints! — Method

Add water target constraints for PowerSystems.HydroReservoir. Only supported for HEAD type.

PowerSimulations.add_feedforward_constraints! — Method

    add_feedforward_constraints(container::PSI.OptimizationContainer,
                    model::PSI.DeviceModel,
                    devices,
                    ff::ReservoirLimitFeedforward)

Constructs a parameterized integral limit constraint to implement feedforward from other models. The Parameters are initialized using the upper boundary values of the provided variables.

sum(variable[var_name, t] for t in 1:affected_periods)/affected_periods <= param_reference[var_name]

LaTeX

$\sum_{t} x \leq param^{max}$

Arguments

container::PowerSimulations.OptimizationContainer : the optimization_container model built in PowerSimulations

model::PowerSimulations.DeviceModel : the device model

devices::IS.FlattenIteratorWrapper{T} : list of devices

ff::ReservoirLimitFeedforward : a instance of the Reservoir Limit Feedforward

PowerSimulations.add_feedforward_constraints! — Method

    add_feedforward_constraints(
        container::PSI.OptimizationContainer,
        ::PSI.DeviceModel,
        devices::IS.FlattenIteratorWrapper{T},
        ff::ReservoirTargetFeedforward,
    ) where {T <: PSY.Component}

Constructs a equality constraint to a fix a variable in one model using the variable value from other model results.

variable[var_name, t] + slack[var_name, t] >= param[var_name, t]

LaTeX

$x + slack >= param$

Arguments

container::PSI.OptimizationContainer : the optimization_container model built in PowerSimulations

model::PowerSimulations.DeviceModel : the device model

devices::IS.FlattenIteratorWrapper{T} : list of devices

ff::ReservoirTargetFeedforward : a instance of the ReservoirTarget Feedforward

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentRunOfRiver Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentRunOfRiver Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchRunOfRiverBudget Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchRunOfRiverBudget Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchRunOfRiver Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchRunOfRiver Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroReservoir with HydroEnergyModelReservoir Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroTurbine with HydroTurbineEnergyCommitment Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroTurbine with HydroTurbineEnergyCommitment Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroTurbine with HydroTurbineEnergyDispatch Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroTurbine with HydroTurbineEnergyDispatch Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroTurbine with HydroTurbineWaterLinearCommitment Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroTurbine with HydroWaterFactorModel Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroReservoir with HydroWaterFactorModel Formulation with only Active Power

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen` with PowerSimulations.FixedOutput Formulation

PowerSimulations.get_min_max_limits — Method

Min and max active Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max reactive Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max active Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max active pump Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max reactive Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max active Power Variable limits