PowerSystems.Branch Formulations

Valid DeviceModels for subtypes of Branch include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(TModelHVDCLine, LossLessLine)	TModelHVDCLine	LossLessLine

DeviceModel(PhaseShiftingTransformer, PhaseAngleControl)	PhaseShiftingTransformer	PhaseAngleControl

DeviceModel(TwoTerminalHVDCLine, HVDCTwoTerminalDispatch)	TwoTerminalHVDCLine	HVDCTwoTerminalDispatch

DeviceModel(TwoTerminalVSCDCLine, HVDCTwoTerminalDispatch)	TwoTerminalVSCDCLine	HVDCTwoTerminalDispatch

DeviceModel(TwoTerminalHVDCLine, HVDCTwoTerminalLossless)	TwoTerminalHVDCLine	HVDCTwoTerminalLossless

DeviceModel(TwoTerminalVSCDCLine, HVDCTwoTerminalLossless)	TwoTerminalVSCDCLine	HVDCTwoTerminalLossless

DeviceModel(TwoTerminalHVDCLine, HVDCTwoTerminalUnbounded)	TwoTerminalHVDCLine	HVDCTwoTerminalUnbounded

DeviceModel(TwoTerminalVSCDCLine, HVDCTwoTerminalUnbounded)	TwoTerminalVSCDCLine	HVDCTwoTerminalUnbounded

DeviceModel(Line, StaticBranch)	Line	StaticBranch

DeviceModel(MonitoredLine, StaticBranch)	MonitoredLine	StaticBranch

DeviceModel(PhaseShiftingTransformer, StaticBranch)	PhaseShiftingTransformer	StaticBranch

DeviceModel(TapTransformer, StaticBranch)	TapTransformer	StaticBranch

DeviceModel(Transformer2W, StaticBranch)	Transformer2W	StaticBranch

DeviceModel(TwoTerminalHVDCLine, StaticBranch)	TwoTerminalHVDCLine	StaticBranch

DeviceModel(TwoTerminalVSCDCLine, StaticBranch)	TwoTerminalVSCDCLine	StaticBranch

DeviceModel(Line, StaticBranchBounds)	Line	StaticBranchBounds

DeviceModel(MonitoredLine, StaticBranchBounds)	MonitoredLine	StaticBranchBounds

DeviceModel(PhaseShiftingTransformer, StaticBranchBounds)	PhaseShiftingTransformer	StaticBranchBounds

DeviceModel(TapTransformer, StaticBranchBounds)	TapTransformer	StaticBranchBounds

DeviceModel(Transformer2W, StaticBranchBounds)	Transformer2W	StaticBranchBounds

DeviceModel(TwoTerminalHVDCLine, StaticBranchBounds)	TwoTerminalHVDCLine	StaticBranchBounds

DeviceModel(TwoTerminalVSCDCLine, StaticBranchBounds)	TwoTerminalVSCDCLine	StaticBranchBounds

DeviceModel(Line, StaticBranchUnbounded)	Line	StaticBranchUnbounded

DeviceModel(MonitoredLine, StaticBranchUnbounded)	MonitoredLine	StaticBranchUnbounded

DeviceModel(PhaseShiftingTransformer, StaticBranchUnbounded)	PhaseShiftingTransformer	StaticBranchUnbounded

DeviceModel(TapTransformer, StaticBranchUnbounded)	TapTransformer	StaticBranchUnbounded

DeviceModel(Transformer2W, StaticBranchUnbounded)	Transformer2W	StaticBranchUnbounded

DeviceModel(TwoTerminalHVDCLine, StaticBranchUnbounded)	TwoTerminalHVDCLine	StaticBranchUnbounded

DeviceModel(TwoTerminalVSCDCLine, StaticBranchUnbounded)	TwoTerminalVSCDCLine	StaticBranchUnbounded

StaticBranch

PowerSimulations.StaticBranch — Type

Branch type to add unbounded flow variables and use flow constraints

StaticBranchBounds

PowerSimulations.StaticBranchBounds — Type

Branch type to add bounded flow variables and use flow constraints

StaticBranchUnbounded

PowerSimulations.StaticBranchUnbounded — Type

Branch type to avoid flow constraints

HVDCTwoTerminalLossless

PowerSimulations.HVDCTwoTerminalLossless — Type

Branch type to represent lossless power flow on DC lines

HVDCTwoTerminalDispatch

PowerSimulations.HVDCTwoTerminalDispatch — Type

Branch type to represent lossy power flow on DC lines

HVDCTwoTerminalUnbounded

PowerSimulations.HVDCTwoTerminalUnbounded — Type

Branch type to avoid flow constraints