ThermalGen Formulations

Valid DeviceModels for subtypes of ThermalGen include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(ThermalMultiStart, FixedOutput)	ThermalMultiStart	FixedOutput

DeviceModel(ThermalStandard, FixedOutput)	ThermalStandard	FixedOutput

DeviceModel(ThermalMultiStart, ThermalBasicDispatch)	ThermalMultiStart	ThermalBasicDispatch

DeviceModel(ThermalStandard, ThermalBasicDispatch)	ThermalStandard	ThermalBasicDispatch

DeviceModel(ThermalMultiStart, ThermalCompactDispatch)	ThermalMultiStart	ThermalCompactDispatch

DeviceModel(ThermalStandard, ThermalCompactDispatch)	ThermalStandard	ThermalCompactDispatch

DeviceModel(ThermalMultiStart, ThermalDispatchNoMin)	ThermalMultiStart	ThermalDispatchNoMin

DeviceModel(ThermalStandard, ThermalDispatchNoMin)	ThermalStandard	ThermalDispatchNoMin

DeviceModel(ThermalMultiStart, ThermalStandardDispatch)	ThermalMultiStart	ThermalStandardDispatch

DeviceModel(ThermalStandard, ThermalStandardDispatch)	ThermalStandard	ThermalStandardDispatch

DeviceModel(ThermalMultiStart, ThermalBasicCompactUnitCommitment)	ThermalMultiStart	ThermalBasicCompactUnitCommitment

DeviceModel(ThermalStandard, ThermalBasicCompactUnitCommitment)	ThermalStandard	ThermalBasicCompactUnitCommitment

DeviceModel(ThermalMultiStart, ThermalCompactUnitCommitment)	ThermalMultiStart	ThermalCompactUnitCommitment

DeviceModel(ThermalStandard, ThermalCompactUnitCommitment)	ThermalStandard	ThermalCompactUnitCommitment

DeviceModel(ThermalMultiStart, ThermalMultiStartUnitCommitment)	ThermalMultiStart	ThermalMultiStartUnitCommitment

DeviceModel(ThermalMultiStart, ThermalBasicUnitCommitment)	ThermalMultiStart	ThermalBasicUnitCommitment

DeviceModel(ThermalStandard, ThermalBasicUnitCommitment)	ThermalStandard	ThermalBasicUnitCommitment

DeviceModel(ThermalMultiStart, ThermalStandardUnitCommitment)	ThermalMultiStart	ThermalStandardUnitCommitment

DeviceModel(ThermalStandard, ThermalStandardUnitCommitment)	ThermalStandard	ThermalStandardUnitCommitment

ThermalBasicDispatch

PowerSimulations.ThermalBasicDispatch — Type

Formulation type to enable basic dispatch without any intertemporal (ramp) constraints

TODO

ThermalCompactDispatch

PowerSimulations.ThermalCompactDispatch — Type

Formulation type to enable thermal compact dispatch

TODO

ThermalDispatchNoMin

PowerSimulations.ThermalDispatchNoMin — Type

Formulation type to enable basic dispatch without any intertemporal constraints and relaxed minimum generation. may not work with PWL cost definitions

TODO

ThermalStandardDispatch

PowerSimulations.ThermalStandardDispatch — Type

Formulation type to enable standard dispatch with a range and enforce intertemporal ramp constraints

TODO

ThermalBasicCompactUnitCommitment

PowerSimulations.ThermalBasicCompactUnitCommitment — Type

Formulation type to enable thermal compact commitment without intertemporal (ramp, min on/off time) constraints

TODO

ThermalCompactUnitCommitment

PowerSimulations.ThermalCompactUnitCommitment — Type

Formulation type to enable thermal compact commitment

TODO

ThermalMultiStartUnitCommitment

PowerSimulations.ThermalMultiStartUnitCommitment — Type

Formulation type to enable pg-lib commitment formulation with startup/shutdown profiles

TODO

ThermalBasicUnitCommitment

PowerSimulations.ThermalBasicUnitCommitment — Type

Formulation type to enable basic unit commitment representation without any intertemporal (ramp, min on/off time) constraints

TODO

ThermalStandardUnitCommitment

PowerSimulations.ThermalStandardUnitCommitment — Type

Formulaiton type to enable standard unit commitment with intertemporal constraints and simplified startup profiles

TODO