BatteryEMS

PowerSystems.BatteryEMS — Type

mutable struct BatteryEMS <: Storage
    name::String
    available::Bool
    bus::ACBus
    prime_mover_type::PrimeMovers
    initial_energy::Float64
    state_of_charge_limits::MinMax
    rating::Float64
    active_power::Float64
    input_active_power_limits::MinMax
    output_active_power_limits::MinMax
    efficiency::NamedTuple{(:in, :out), Tuple{Float64, Float64}}
    reactive_power::Float64
    reactive_power_limits::Union{Nothing, MinMax}
    base_power::Float64
    operation_cost::StorageManagementCost
    storage_target::Float64
    cycle_limits::Int
    services::Vector{Service}
    dynamic_injector::Union{Nothing, DynamicInjection}
    ext::Dict{String, Any}
    time_series_container::InfrastructureSystems.TimeSeriesContainer
    internal::InfrastructureSystemsInternal
end

Data structure for a battery compatible with energy management formulations.

Arguments

name::String

available::Bool

bus::ACBus

prime_mover_type::PrimeMovers: Prime mover technology according to EIA 923

initial_energy::Float64: State of Charge of the Battery p.u.-hr, validation range: (0, nothing), action if invalid: error

state_of_charge_limits::MinMax: Maximum and Minimum storage capacity in p.u.-hr, validation range: (0, nothing), action if invalid: error

rating::Float64

active_power::Float64

input_active_power_limits::MinMax, validation range: (0, nothing), action if invalid: error

output_active_power_limits::MinMax, validation range: (0, nothing), action if invalid: error

efficiency::NamedTuple{(:in, :out), Tuple{Float64, Float64}}, validation range: (0, 1), action if invalid: warn

reactive_power::Float64, validation range: reactive_power_limits, action if invalid: warn

reactive_power_limits::Union{Nothing, MinMax}

base_power::Float64: Base power of the unit in MVA, validation range: (0, nothing), action if invalid: warn

operation_cost::StorageManagementCost

storage_target::Float64: Storage target at the end of simulation as ratio of storage capacity.

cycle_limits::Int: Storage Maximum number of cycles per year

services::Vector{Service}: Services that this device contributes to

dynamic_injector::Union{Nothing, DynamicInjection}: corresponding dynamic injection device

ext::Dict{String, Any}

time_series_container::InfrastructureSystems.TimeSeriesContainer: internal time_series storage

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

InfrastructureSystems.get_name — Method

get_name(value::BatteryEMS) -> String

Get BatteryEMSname.

InfrastructureSystems.get_time_series_container — Method

get_time_series_container(
    value::BatteryEMS
) -> InfrastructureSystems.TimeSeriesContainer

Get BatteryEMStime_series_container.

PowerSystems.get_active_power — Method

get_active_power(value::BatteryEMS) -> Any

Get BatteryEMSactive_power.

PowerSystems.get_available — Method

get_available(value::BatteryEMS) -> Bool

Get BatteryEMSavailable.

PowerSystems.get_base_power — Method

get_base_power(value::BatteryEMS) -> Float64

Get BatteryEMSbase_power.

PowerSystems.get_bus — Method

get_bus(value::BatteryEMS) -> ACBus

Get BatteryEMSbus.

PowerSystems.get_cycle_limits — Method

get_cycle_limits(value::BatteryEMS) -> Int64

Get BatteryEMScycle_limits.

PowerSystems.get_dynamic_injector — Method

get_dynamic_injector(
    value::BatteryEMS
) -> Union{Nothing, DynamicInjection}

Get BatteryEMSdynamic_injector.

PowerSystems.get_efficiency — Method

get_efficiency(
    value::BatteryEMS
) -> @NamedTuple{in::Float64, out::Float64}

Get BatteryEMSefficiency.

PowerSystems.get_ext — Method

get_ext(value::BatteryEMS) -> Dict{String, Any}

Get BatteryEMSext.

PowerSystems.get_initial_energy — Method

get_initial_energy(value::BatteryEMS) -> Any

Get BatteryEMSinitial_energy.

PowerSystems.get_input_active_power_limits — Method

get_input_active_power_limits(
    value::BatteryEMS
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get BatteryEMSinput_active_power_limits.

PowerSystems.get_operation_cost — Method

get_operation_cost(
    value::BatteryEMS
) -> StorageManagementCost

Get BatteryEMSoperation_cost.

PowerSystems.get_output_active_power_limits — Method

get_output_active_power_limits(
    value::BatteryEMS
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get BatteryEMSoutput_active_power_limits.

PowerSystems.get_prime_mover_type — Method

get_prime_mover_type(value::BatteryEMS) -> PrimeMovers

Get BatteryEMSprime_mover_type.

PowerSystems.get_rating — Method

get_rating(value::BatteryEMS) -> Any

Get BatteryEMSrating.

PowerSystems.get_reactive_power — Method

get_reactive_power(value::BatteryEMS) -> Any

Get BatteryEMSreactive_power.

PowerSystems.get_reactive_power_limits — Method

get_reactive_power_limits(
    value::BatteryEMS
) -> Union{Nothing, NamedTuple{(:min, :max), <:Tuple{Any, Any}}}

Get BatteryEMSreactive_power_limits.

PowerSystems.get_services — Method

get_services(value::BatteryEMS) -> Vector{Service}

Get BatteryEMSservices.

PowerSystems.get_state_of_charge_limits — Method

get_state_of_charge_limits(
    value::BatteryEMS
) -> NamedTuple{(:min, :max), <:Tuple{Any, Any}}

Get BatteryEMSstate_of_charge_limits.

PowerSystems.get_storage_target — Method

get_storage_target(value::BatteryEMS) -> Float64

Get BatteryEMSstorage_target.

PowerSystems.set_active_power! — Method

set_active_power!(value::BatteryEMS, val) -> Any

Set BatteryEMSactive_power.

PowerSystems.set_available! — Method

set_available!(value::BatteryEMS, val) -> Any

Set BatteryEMSavailable.

PowerSystems.set_base_power! — Method

set_base_power!(value::BatteryEMS, val) -> Any

Set BatteryEMSbase_power.

PowerSystems.set_bus! — Method

set_bus!(value::BatteryEMS, val) -> Any

Set BatteryEMSbus.

PowerSystems.set_cycle_limits! — Method

set_cycle_limits!(value::BatteryEMS, val) -> Any

Set BatteryEMScycle_limits.

PowerSystems.set_efficiency! — Method

set_efficiency!(value::BatteryEMS, val) -> Any

Set BatteryEMSefficiency.

PowerSystems.set_ext! — Method

set_ext!(value::BatteryEMS, val) -> Any

Set BatteryEMSext.

PowerSystems.set_initial_energy! — Method

set_initial_energy!(value::BatteryEMS, val) -> Any

Set BatteryEMSinitial_energy.

PowerSystems.set_input_active_power_limits! — Method

set_input_active_power_limits!(
    value::BatteryEMS,
    val
) -> Any

Set BatteryEMSinput_active_power_limits.

PowerSystems.set_operation_cost! — Method

set_operation_cost!(value::BatteryEMS, val) -> Any

Set BatteryEMSoperation_cost.

PowerSystems.set_output_active_power_limits! — Method

set_output_active_power_limits!(
    value::BatteryEMS,
    val
) -> Any

Set BatteryEMSoutput_active_power_limits.

PowerSystems.set_prime_mover_type! — Method

set_prime_mover_type!(value::BatteryEMS, val) -> Any

Set BatteryEMSprime_mover_type.

PowerSystems.set_rating! — Method

set_rating!(value::BatteryEMS, val) -> Any

Set BatteryEMSrating.

PowerSystems.set_reactive_power! — Method

set_reactive_power!(value::BatteryEMS, val) -> Any

Set BatteryEMSreactive_power.

PowerSystems.set_reactive_power_limits! — Method

set_reactive_power_limits!(value::BatteryEMS, val) -> Any

Set BatteryEMSreactive_power_limits.

PowerSystems.set_services! — Method

set_services!(value::BatteryEMS, val) -> Any

Set BatteryEMSservices.

PowerSystems.set_state_of_charge_limits! — Method

set_state_of_charge_limits!(value::BatteryEMS, val) -> Any

Set BatteryEMSstate_of_charge_limits.

PowerSystems.set_storage_target! — Method

set_storage_target!(value::BatteryEMS, val) -> Any

Set BatteryEMSstorage_target.

PowerSystems.set_time_series_container! — Method

set_time_series_container!(value::BatteryEMS, val) -> Any

Set BatteryEMStime_series_container.