Machine

AndersonFouadMachine

PowerSystems.AndersonFouadMachine — Type

mutable struct AndersonFouadMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 6-states synchronous machine: Anderson-Fouad model

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

ψq: q-axis stator flux,
ψd: d-axis stator flux,
eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: AndersonFouadMachine has 6 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineR.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineTd0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineTd0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineTq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineTq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineXd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineXd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineXq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineXq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::AndersonFouadMachine) -> Float64

Get AndersonFouadMachineXq_pp.

PowerSystems.get_ext — Method

get_ext(value::AndersonFouadMachine) -> Dict{String, Any}

Get AndersonFouadMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::AndersonFouadMachine) -> Int64

Get AndersonFouadMachinen_states.

PowerSystems.get_states — Method

get_states(value::AndersonFouadMachine) -> Vector{Symbol}

Get AndersonFouadMachinestates.

PowerSystems.set_R! — Method

set_R!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineR.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineTd0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineTd0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineTq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineTq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineXd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineXd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineXq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineXq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineXq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::AndersonFouadMachine, val) -> Any

Set AndersonFouadMachineext.

BaseMachine

PowerSystems.BaseMachine — Type

mutable struct BaseMachine <: Machine
    R::Float64
    Xd_p::Float64
    eq_p::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of a Classic Machine: GENCLS in PSSE and PSLF

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd_p::Float64: Reactance after EMF in machine per unit, validation range: (0, nothing)

eq_p::Float64: Fixed EMF behind the impedance, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: BaseMachine has no states

n_states::Int: BaseMachine has no states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::BaseMachine) -> Float64

Get BaseMachineR.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::BaseMachine) -> Float64

Get BaseMachineXd_p.

PowerSystems.get_eq_p — Method

get_eq_p(value::BaseMachine) -> Float64

Get BaseMachineeq_p.

PowerSystems.get_ext — Method

get_ext(value::BaseMachine) -> Dict{String, Any}

Get BaseMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::BaseMachine) -> Int64

Get BaseMachinen_states.

PowerSystems.get_states — Method

get_states(value::BaseMachine) -> Vector{Symbol}

Get BaseMachinestates.

PowerSystems.set_R! — Method

set_R!(value::BaseMachine, val) -> Any

Set BaseMachineR.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::BaseMachine, val) -> Any

Set BaseMachineXd_p.

PowerSystems.set_eq_p! — Method

set_eq_p!(value::BaseMachine, val) -> Any

Set BaseMachineeq_p.

PowerSystems.set_ext! — Method

set_ext!(value::BaseMachine, val) -> Any

Set BaseMachineext.

FullMachine

PowerSystems.FullMachine — Type

mutable struct FullMachine <: Machine
    R::Float64
    R_f::Float64
    R_1d::Float64
    R_1q::Float64
    L_d::Float64
    L_q::Float64
    L_ad::Float64
    L_aq::Float64
    L_f1d::Float64
    L_ff::Float64
    L_1d::Float64
    L_1q::Float64
    ext::Dict{String, Any}
    inv_d_fluxlink::Array{Float64,2}
    inv_q_fluxlink::Array{Float64,2}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameter of a full order flux stator-rotor model without zero sequence flux in the stator. The derivative of stator fluxes (ψd and ψq) is NOT neglected. Only one q-axis damping circuit is considered. All parameters are in machine per unit. Refer to Chapter 3 of Power System Stability and Control by P. Kundur or Chapter 11 of Power System Dynamics: Stability and Control, by J. Machowski, J. Bialek and J. Bumby, for more details. Note that the models are somewhat different (but equivalent) due to the different Park Transformation used in both books.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

R_f::Float64: Field rotor winding resistance in per unit, validation range: (0, nothing)

R_1d::Float64: Damping rotor winding resistance on d-axis in per unit. This value is denoted as RD in Machowski., validation range: (0, nothing)

R_1q::Float64: Damping rotor winding resistance on q-axis in per unit. This value is denoted as RQ in Machowski., validation range: (0, nothing)

L_d::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the d-axis of the rotor, in per unit. This value is denoted as Lad + Ll in Kundur (and Ld in Machowski)., validation range: (0, nothing)

L_q::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the q-axis of the rotor, in per unit. This value is denoted as Laq + Ll in Kundur., validation range: (0, nothing)

L_ad::Float64: Mutual inductance between stator winding and rotor field (and damping) winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_aq::Float64: Mutual inductance between stator winding and rotor damping winding inductance on q-axis, in per unit, validation range: (0, nothing)

L_f1d::Float64: Mutual inductance between rotor field winding and rotor damping winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_ff::Float64: Field rotor winding inductance, in per unit, validation range: (0, nothing)

L_1d::Float64: Inductance of the d-axis rotor damping circuit, in per unit, validation range: (0, nothing)

L_1q::Float64: Inductance of the q-axis rotor damping circuit, in per unit, validation range: (0, nothing)

ext::Dict{String, Any}

inv_d_fluxlink::Array{Float64,2}: Equations 3.127, 3.130, 3.131 From Kundur

inv_q_fluxlink::Array{Float64,2}: Equations 3.128, 3.132 From Kundur

states::Vector{Symbol}: The states are:

ψd: d-axis stator flux,
ψq: q-axis stator flux,
ψf: field rotor flux,
ψ1d: d-axis rotor damping flux,
ψ1q: q-axis rotor damping flux

n_states::Int: FullMachine has 5 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_L_1d — Method

get_L_1d(value::FullMachine) -> Float64

Get FullMachineL_1d.

PowerSystems.get_L_1q — Method

get_L_1q(value::FullMachine) -> Float64

Get FullMachineL_1q.

PowerSystems.get_L_ad — Method

get_L_ad(value::FullMachine) -> Float64

Get FullMachineL_ad.

PowerSystems.get_L_aq — Method

get_L_aq(value::FullMachine) -> Float64

Get FullMachineL_aq.

PowerSystems.get_L_d — Method

get_L_d(value::FullMachine) -> Float64

Get FullMachineL_d.

PowerSystems.get_L_f1d — Method

get_L_f1d(value::FullMachine) -> Float64

Get FullMachineL_f1d.

PowerSystems.get_L_ff — Method

get_L_ff(value::FullMachine) -> Float64

Get FullMachineL_ff.

PowerSystems.get_L_q — Method

get_L_q(value::FullMachine) -> Float64

Get FullMachineL_q.

PowerSystems.get_R — Method

get_R(value::FullMachine) -> Float64

Get FullMachineR.

PowerSystems.get_R_1d — Method

get_R_1d(value::FullMachine) -> Float64

Get FullMachineR_1d.

PowerSystems.get_R_1q — Method

get_R_1q(value::FullMachine) -> Float64

Get FullMachineR_1q.

PowerSystems.get_R_f — Method

get_R_f(value::FullMachine) -> Float64

Get FullMachineR_f.

PowerSystems.get_ext — Method

get_ext(value::FullMachine) -> Dict{String, Any}

Get FullMachineext.

PowerSystems.get_inv_d_fluxlink — Method

get_inv_d_fluxlink(value::FullMachine) -> Matrix{Float64}

Get FullMachineinv_d_fluxlink.

PowerSystems.get_inv_q_fluxlink — Method

get_inv_q_fluxlink(value::FullMachine) -> Matrix{Float64}

Get FullMachineinv_q_fluxlink.

PowerSystems.get_n_states — Method

get_n_states(value::FullMachine) -> Int64

Get FullMachinen_states.

PowerSystems.get_states — Method

get_states(value::FullMachine) -> Vector{Symbol}

Get FullMachinestates.

PowerSystems.set_L_1d! — Method

set_L_1d!(value::FullMachine, val) -> Any

Set FullMachineL_1d.

PowerSystems.set_L_1q! — Method

set_L_1q!(value::FullMachine, val) -> Any

Set FullMachineL_1q.

PowerSystems.set_L_ad! — Method

set_L_ad!(value::FullMachine, val) -> Any

Set FullMachineL_ad.

PowerSystems.set_L_aq! — Method

set_L_aq!(value::FullMachine, val) -> Any

Set FullMachineL_aq.

PowerSystems.set_L_d! — Method

set_L_d!(value::FullMachine, val) -> Any

Set FullMachineL_d.

PowerSystems.set_L_f1d! — Method

set_L_f1d!(value::FullMachine, val) -> Any

Set FullMachineL_f1d.

PowerSystems.set_L_ff! — Method

set_L_ff!(value::FullMachine, val) -> Any

Set FullMachineL_ff.

PowerSystems.set_L_q! — Method

set_L_q!(value::FullMachine, val) -> Any

Set FullMachineL_q.

PowerSystems.set_R! — Method

set_R!(value::FullMachine, val) -> Any

Set FullMachineR.

PowerSystems.set_R_1d! — Method

set_R_1d!(value::FullMachine, val) -> Any

Set FullMachineR_1d.

PowerSystems.set_R_1q! — Method

set_R_1q!(value::FullMachine, val) -> Any

Set FullMachineR_1q.

PowerSystems.set_R_f! — Method

set_R_f!(value::FullMachine, val) -> Any

Set FullMachineR_f.

PowerSystems.set_ext! — Method

set_ext!(value::FullMachine, val) -> Any

Set FullMachineext.

PowerSystems.set_inv_d_fluxlink! — Method

set_inv_d_fluxlink!(value::FullMachine, val) -> Any

Set FullMachineinv_d_fluxlink.

PowerSystems.set_inv_q_fluxlink! — Method

set_inv_q_fluxlink!(value::FullMachine, val) -> Any

Set FullMachineinv_q_fluxlink.

MarconatoMachine

PowerSystems.MarconatoMachine — Type

mutable struct MarconatoMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    T_AA::Float64
    ext::Dict{String, Any}
    γd::Float64
    γq::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 6-states synchronous machine: Marconato model

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

T_AA::Float64: Time constant of d-axis additional leakage, validation range: (0, nothing)

ext::Dict{String, Any}

γd::Float64

γq::Float64

states::Vector{Symbol}: The states are:

ψq: q-axis stator flux,
ψd: d-axis stator flux,
eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: MarconatoMachine has 6 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::MarconatoMachine) -> Float64

Get MarconatoMachineR.

PowerSystems.get_T_AA — Method

get_T_AA(value::MarconatoMachine) -> Float64

Get MarconatoMachineT_AA.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::MarconatoMachine) -> Float64

Get MarconatoMachineTd0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachineTd0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::MarconatoMachine) -> Float64

Get MarconatoMachineTq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachineTq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::MarconatoMachine) -> Float64

Get MarconatoMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::MarconatoMachine) -> Float64

Get MarconatoMachineXd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachineXd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::MarconatoMachine) -> Float64

Get MarconatoMachineXq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::MarconatoMachine) -> Float64

Get MarconatoMachineXq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::MarconatoMachine) -> Float64

Get MarconatoMachineXq_pp.

PowerSystems.get_ext — Method

get_ext(value::MarconatoMachine) -> Dict{String, Any}

Get MarconatoMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::MarconatoMachine) -> Int64

Get MarconatoMachinen_states.

PowerSystems.get_states — Method

get_states(value::MarconatoMachine) -> Vector{Symbol}

Get MarconatoMachinestates.

PowerSystems.get_γd — Method

get_γd(value::MarconatoMachine) -> Float64

Get MarconatoMachineγd.

PowerSystems.get_γq — Method

get_γq(value::MarconatoMachine) -> Float64

Get MarconatoMachineγq.

PowerSystems.set_R! — Method

set_R!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineR.

PowerSystems.set_T_AA! — Method

set_T_AA!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineT_AA.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineTd0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineTd0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineTq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineTq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineXd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineXd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineXq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineXq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineXq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineext.

PowerSystems.set_γd! — Method

set_γd!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineγd.

PowerSystems.set_γq! — Method

set_γq!(value::MarconatoMachine, val) -> Any

Set MarconatoMachineγq.

OneDOneQMachine

PowerSystems.OneDOneQMachine — Type

mutable struct OneDOneQMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Td0_p::Float64
    Tq0_p::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states synchronous machine: Simplified Marconato model The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage

n_states::Int: OneDOneQMachine has 2 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::OneDOneQMachine) -> Float64

Get OneDOneQMachineR.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachineTd0_p.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachineTq0_p.

PowerSystems.get_Xd — Method

get_Xd(value::OneDOneQMachine) -> Float64

Get OneDOneQMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachineXd_p.

PowerSystems.get_Xq — Method

get_Xq(value::OneDOneQMachine) -> Float64

Get OneDOneQMachineXq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::OneDOneQMachine) -> Float64

Get OneDOneQMachineXq_p.

PowerSystems.get_ext — Method

get_ext(value::OneDOneQMachine) -> Dict{String, Any}

Get OneDOneQMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::OneDOneQMachine) -> Int64

Get OneDOneQMachinen_states.

PowerSystems.get_states — Method

get_states(value::OneDOneQMachine) -> Vector{Symbol}

Get OneDOneQMachinestates.

PowerSystems.set_R! — Method

set_R!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineR.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineTd0_p.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineTq0_p.

PowerSystems.set_Xd! — Method

set_Xd!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineXd_p.

PowerSystems.set_Xq! — Method

set_Xq!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineXq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineXq_p.

PowerSystems.set_ext! — Method

set_ext!(value::OneDOneQMachine, val) -> Any

Set OneDOneQMachineext.

RoundRotorExponential

PowerSystems.RoundRotorExponential — Type

mutable struct RoundRotorExponential <: Machine
    base_machine::RoundRotorMachine
    saturation_coeffs::Tuple{Float64, Float64}

4-states round-rotor synchronous machine with exponential saturation: IEEE Std 1110 5.3.2 (Model 2.2). GENROE model in PSSE and PSLF.

Arguments

base_machine::RoundRotorMachine: Round Rotor Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for exponential model.

RoundRotorMachine

PowerSystems.RoundRotorMachine — Type

mutable struct RoundRotorMachine <: Machine
    R::Float64
    Td0_p::Float64
    Td0_pp::Float64
    Tq0_p::Float64
    Tq0_pp::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xl::Float64
    Se::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    γ_d1::Float64
    γ_q1::Float64
    γ_d2::Float64
    γ_q2::Float64
    γ_qd::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states round-rotor synchronous machine with quadratic/exponential saturation: IEEE Std 1110 5.3.2 (Model 2.2). GENROU or GENROE model in PSSE and PSLF.

Arguments

R::Float64: Armature resistance, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit. Note: Xdpp = Xqpp, validation range: (0, nothing)

Xl::Float64: Stator leakage reactance, validation range: (0, nothing)

Se::Tuple{Float64, Float64}: Saturation factor at 1 and 1.2 pu flux: S(1.0) = B(|ψ_pp|-A)^2

ext::Dict{String, Any}

γ_d1::Float64: γ_d1 parameter

γ_q1::Float64: γ_q1 parameter

γ_d2::Float64: γ_d2 parameter

γ_q2::Float64: γ_q2 parameter

γ_qd::Float64: γ_qd parameter

states::Vector{Symbol}: The states are:

eq_p: q-axis generator voltage behind the transient reactance,
ed_p: d-axis generator voltage behind the transient reactance,
ψ_kd: flux linkage in the first equivalent damping circuit in the d-axis,
ψ_kq: flux linkage in the first equivalent damping circuit in the d-axis

n_states::Int: RoundRotorMachine has 4 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineR.

PowerSystems.get_Se — Method

get_Se(value::RoundRotorMachine) -> Tuple{Float64, Float64}

Get RoundRotorMachineSe.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineTd0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineTd0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineTq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineTq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineXd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineXd_pp.

PowerSystems.get_Xl — Method

get_Xl(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineXl.

PowerSystems.get_Xq — Method

get_Xq(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineXq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineXq_p.

PowerSystems.get_ext — Method

get_ext(value::RoundRotorMachine) -> Dict{String, Any}

Get RoundRotorMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::RoundRotorMachine) -> Int64

Get RoundRotorMachinen_states.

PowerSystems.get_states — Method

get_states(value::RoundRotorMachine) -> Vector{Symbol}

Get RoundRotorMachinestates.

PowerSystems.get_γ_d1 — Method

get_γ_d1(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineγ_d1.

PowerSystems.get_γ_d2 — Method

get_γ_d2(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineγ_d2.

PowerSystems.get_γ_q1 — Method

get_γ_q1(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineγ_q1.

PowerSystems.get_γ_q2 — Method

get_γ_q2(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineγ_q2.

PowerSystems.get_γ_qd — Method

get_γ_qd(value::RoundRotorMachine) -> Float64

Get RoundRotorMachineγ_qd.

PowerSystems.set_R! — Method

set_R!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineR.

PowerSystems.set_Se! — Method

set_Se!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineSe.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineTd0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineTd0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineTq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineTq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineXd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineXd_pp.

PowerSystems.set_Xl! — Method

set_Xl!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineXl.

PowerSystems.set_Xq! — Method

set_Xq!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineXq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineXq_p.

PowerSystems.set_ext! — Method

set_ext!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineext.

PowerSystems.set_γ_d1! — Method

set_γ_d1!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineγ_d1.

PowerSystems.set_γ_d2! — Method

set_γ_d2!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineγ_d2.

PowerSystems.set_γ_q1! — Method

set_γ_q1!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineγ_q1.

PowerSystems.set_γ_q2! — Method

set_γ_q2!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineγ_q2.

PowerSystems.set_γ_qd! — Method

set_γ_qd!(value::RoundRotorMachine, val) -> Any

Set RoundRotorMachineγ_qd.

RoundRotorQuadratic

PowerSystems.RoundRotorQuadratic — Type

mutable struct RoundRotorQuadratic <: Machine
    base_machine::RoundRotorMachine
    saturation_coeffs::Tuple{Float64, Float64}

4-states round-rotor synchronous machine with quadratic saturation: IEEE Std 1110 5.3.2 (Model 2.2). GENROU model in PSSE and PSLF.

Arguments

base_machine::RoundRotorMachine: Round Rotor Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for quadratic model.

SalientPoleExponential

PowerSystems.SalientPoleExponential — Type

mutable struct SalientPoleExponential <: Machine
    base_machine::SalientPoleMachine
    saturation_coeffs::Tuple{Float64, Float64}

3-states salient-pole synchronous machine with exponential saturation: IEEE Std 1110 5.3.2 (Model 2.1). GENSAE in PSSE and PSLF.

Arguments:

base_machine::SalientPoleMachine: Salient Pole Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for exponential model.

SalientPoleMachine

PowerSystems.SalientPoleMachine — Type

mutable struct SalientPoleMachine <: Machine
    R::Float64
    Td0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xd_pp::Float64
    Xl::Float64
    Se::Tuple{Float64, Float64}
    ext::Dict{String, Any}
    γ_d1::Float64
    γ_q1::Float64
    γ_d2::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 3-states salient-pole synchronous machine with quadratic/exponential saturation: IEEE Std 1110 5.3.1 (Model 2.1). GENSAL or GENSAE model in PSSE and PSLF.

Arguments

R::Float64: Armature resistance, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit. Note: Xdpp = Xqpp, validation range: (0, nothing)

Xl::Float64: Stator leakage reactance, validation range: (0, nothing)

Se::Tuple{Float64, Float64}: Saturation factor at 1 and 1.2 pu flux: Se(eqp) = B(eqp-A)^2

ext::Dict{String, Any}

γ_d1::Float64: γ_d1 parameter

γ_q1::Float64: γ_q1 parameter

γ_d2::Float64: γ_d2 parameter

states::Vector{Symbol}: The states are:

eq_p: q-axis generator voltage behind the transient reactance,
ψ_kd: flux linkage in the first equivalent damping circuit in the d-axis,
ψq_pp: phasonf of the subtransient flux linkage in the q-axis

n_states::Int: SalientPoleMachine has 3 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineR.

PowerSystems.get_Se — Method

get_Se(value::SalientPoleMachine) -> Tuple{Float64, Float64}

Get SalientPoleMachineSe.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineTd0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineTd0_pp.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineTq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineXd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineXd_pp.

PowerSystems.get_Xl — Method

get_Xl(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineXl.

PowerSystems.get_Xq — Method

get_Xq(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineXq.

PowerSystems.get_ext — Method

get_ext(value::SalientPoleMachine) -> Dict{String, Any}

Get SalientPoleMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::SalientPoleMachine) -> Int64

Get SalientPoleMachinen_states.

PowerSystems.get_states — Method

get_states(value::SalientPoleMachine) -> Vector{Symbol}

Get SalientPoleMachinestates.

PowerSystems.get_γ_d1 — Method

get_γ_d1(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineγ_d1.

PowerSystems.get_γ_d2 — Method

get_γ_d2(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineγ_d2.

PowerSystems.get_γ_q1 — Method

get_γ_q1(value::SalientPoleMachine) -> Float64

Get SalientPoleMachineγ_q1.

PowerSystems.set_R! — Method

set_R!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineR.

PowerSystems.set_Se! — Method

set_Se!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineSe.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineTd0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineTd0_pp.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineTq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineXd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineXd_pp.

PowerSystems.set_Xl! — Method

set_Xl!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineXl.

PowerSystems.set_Xq! — Method

set_Xq!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineXq.

PowerSystems.set_ext! — Method

set_ext!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineext.

PowerSystems.set_γ_d1! — Method

set_γ_d1!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineγ_d1.

PowerSystems.set_γ_d2! — Method

set_γ_d2!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineγ_d2.

PowerSystems.set_γ_q1! — Method

set_γ_q1!(value::SalientPoleMachine, val) -> Any

Set SalientPoleMachineγ_q1.

SalientPoleQuadratic

PowerSystems.SalientPoleQuadratic — Type

mutable struct SalientPoleQuadratic <: Machine
    base_machine::SalientPoleMachine
    saturation_coeffs::Tuple{Float64, Float64}

3-states salient-pole synchronous machine with exponential saturation: IEEE Std 1110 5.3.2 (Model 2.1). GENSAL in PSSE and PSLF.

Arguments:

base_machine::SalientPoleMachine: Salient Pole Machine model.

saturation_coeffs::Tuple{Float64, Float64}`: Saturation coefficients for quadratic model.

SauerPaiMachine

PowerSystems.SauerPaiMachine — Type

mutable struct SauerPaiMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Xl::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    ext::Dict{String, Any}
    γ_d1::Float64
    γ_q1::Float64
    γ_d2::Float64
    γ_q2::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of synchronous machine: Sauer Pai model

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xl::Float64: Stator Leakage Reactance, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

ext::Dict{String, Any}

γ_d1::Float64

γ_q1::Float64

γ_d2::Float64

γ_q2::Float64

states::Vector{Symbol}: The states are:

ψq: q-axis stator flux,
ψd: d-axis stator flux,
eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage
ψd_pp: subtransient flux linkage in the d-axis
ψq_pp: subtransient flux linkage in the q-axis

n_states::Int: SauerPaiMachine has 6 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineR.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineTd0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineTd0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineTq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineTq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineXd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineXd_pp.

PowerSystems.get_Xl — Method

get_Xl(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineXl.

PowerSystems.get_Xq — Method

get_Xq(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineXq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineXq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineXq_pp.

PowerSystems.get_ext — Method

get_ext(value::SauerPaiMachine) -> Dict{String, Any}

Get SauerPaiMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::SauerPaiMachine) -> Int64

Get SauerPaiMachinen_states.

PowerSystems.get_states — Method

get_states(value::SauerPaiMachine) -> Vector{Symbol}

Get SauerPaiMachinestates.

PowerSystems.get_γ_d1 — Method

get_γ_d1(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineγ_d1.

PowerSystems.get_γ_d2 — Method

get_γ_d2(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineγ_d2.

PowerSystems.get_γ_q1 — Method

get_γ_q1(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineγ_q1.

PowerSystems.get_γ_q2 — Method

get_γ_q2(value::SauerPaiMachine) -> Float64

Get SauerPaiMachineγ_q2.

PowerSystems.set_R! — Method

set_R!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineR.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineTd0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineTd0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineTq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineTq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineXd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineXd_pp.

PowerSystems.set_Xl! — Method

set_Xl!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineXl.

PowerSystems.set_Xq! — Method

set_Xq!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineXq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineXq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineXq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineext.

PowerSystems.set_γ_d1! — Method

set_γ_d1!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineγ_d1.

PowerSystems.set_γ_d2! — Method

set_γ_d2!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineγ_d2.

PowerSystems.set_γ_q1! — Method

set_γ_q1!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineγ_q1.

PowerSystems.set_γ_q2! — Method

set_γ_q2!(value::SauerPaiMachine, val) -> Any

Set SauerPaiMachineγ_q2.

SimpleAFMachine

PowerSystems.SimpleAFMachine — Type

mutable struct SimpleAFMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states simplified Anderson-Fouad (SimpleAFMachine) model. The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected. If transmission dynamics is considered use the full order Anderson Fouad model.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

ext::Dict{String, Any}

states::Vector{Symbol}: The states are:

eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: SimpleAFMachine has 4 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineR.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineTd0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineTd0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineTq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineTq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineXd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineXd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineXq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineXq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::SimpleAFMachine) -> Float64

Get SimpleAFMachineXq_pp.

PowerSystems.get_ext — Method

get_ext(value::SimpleAFMachine) -> Dict{String, Any}

Get SimpleAFMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::SimpleAFMachine) -> Int64

Get SimpleAFMachinen_states.

PowerSystems.get_states — Method

get_states(value::SimpleAFMachine) -> Vector{Symbol}

Get SimpleAFMachinestates.

PowerSystems.set_R! — Method

set_R!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineR.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineTd0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineTd0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineTq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineTq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineXd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineXd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineXq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineXq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineXq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::SimpleAFMachine, val) -> Any

Set SimpleAFMachineext.

SimpleFullMachine

PowerSystems.SimpleFullMachine — Type

mutable struct SimpleFullMachine <: Machine
    R::Float64
    R_f::Float64
    R_1d::Float64
    R_1q::Float64
    L_d::Float64
    L_q::Float64
    L_ad::Float64
    L_aq::Float64
    L_f1d::Float64
    L_ff::Float64
    L_1d::Float64
    L_1q::Float64
    ext::Dict{String, Any}
    inv_d_fluxlink::Array{Float64,2}
    inv_q_fluxlink::Array{Float64,2}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameter of a full order flux stator-rotor model without zero sequence flux in the stator. The derivative of stator fluxes (ψd and ψq) is neglected. This is standard when transmission network dynamics is neglected. Only one q-axis damping circuit is considered. All per unit are in machine per unit. Refer to Chapter 3 of Power System Stability and Control by P. Kundur or Chapter 11 of Power System Dynamics: Stability and Control, by J. Machowski, J. Bialek and J. Bumby, for more details. Note that the models are somewhat different (but equivalent) due to the different Park Transformation used in both books.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

R_f::Float64: Field rotor winding resistance in per unit, validation range: (0, nothing)

R_1d::Float64: Damping rotor winding resistance on d-axis in per unit. This value is denoted as RD in Machowski., validation range: (0, nothing)

R_1q::Float64: Damping rotor winding resistance on q-axis in per unit. This value is denoted as RQ in Machowski., validation range: (0, nothing)

L_d::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the d-axis of the rotor, in per unit. This value is denoted as Lad + Ll in Kundur (and Ld in Machowski)., validation range: (0, nothing)

L_q::Float64: Inductance of fictitious damping that represent the effect of the three-phase stator winding in the q-axis of the rotor, in per unit. This value is denoted as Laq + Ll in Kundur., validation range: (0, nothing)

L_ad::Float64: Mutual inductance between stator winding and rotor field (and damping) winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_aq::Float64: Mutual inductance between stator winding and rotor damping winding inductance on q-axis, in per unit, validation range: (0, nothing)

L_f1d::Float64: Mutual inductance between rotor field winding and rotor damping winding inductance on d-axis, in per unit, validation range: (0, nothing)

L_ff::Float64: Field rotor winding inductance, in per unit, validation range: (0, nothing)

L_1d::Float64: Inductance of the d-axis rotor damping circuit, in per unit, validation range: (0, nothing)

L_1q::Float64: Inductance of the q-axis rotor damping circuit, in per unit, validation range: (0, nothing)

ext::Dict{String, Any}

inv_d_fluxlink::Array{Float64,2}: Equations 3.127, 3.130, 3.131 From Kundur

inv_q_fluxlink::Array{Float64,2}: Equations 3.128, 3.132 From Kundur

states::Vector{Symbol}: The states are:

ψf: field rotor flux,
ψ1d: d-axis rotor damping flux,
ψ1q: q-axis rotor damping flux

n_states::Int: SimpleFullMachine has 3 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_L_1d — Method

get_L_1d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_1d.

PowerSystems.get_L_1q — Method

get_L_1q(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_1q.

PowerSystems.get_L_ad — Method

get_L_ad(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_ad.

PowerSystems.get_L_aq — Method

get_L_aq(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_aq.

PowerSystems.get_L_d — Method

get_L_d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_d.

PowerSystems.get_L_f1d — Method

get_L_f1d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_f1d.

PowerSystems.get_L_ff — Method

get_L_ff(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_ff.

PowerSystems.get_L_q — Method

get_L_q(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineL_q.

PowerSystems.get_R — Method

get_R(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineR.

PowerSystems.get_R_1d — Method

get_R_1d(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineR_1d.

PowerSystems.get_R_1q — Method

get_R_1q(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineR_1q.

PowerSystems.get_R_f — Method

get_R_f(value::SimpleFullMachine) -> Float64

Get SimpleFullMachineR_f.

PowerSystems.get_ext — Method

get_ext(value::SimpleFullMachine) -> Dict{String, Any}

Get SimpleFullMachineext.

PowerSystems.get_inv_d_fluxlink — Method

get_inv_d_fluxlink(
    value::SimpleFullMachine
) -> Matrix{Float64}

Get SimpleFullMachineinv_d_fluxlink.

PowerSystems.get_inv_q_fluxlink — Method

get_inv_q_fluxlink(
    value::SimpleFullMachine
) -> Matrix{Float64}

Get SimpleFullMachineinv_q_fluxlink.

PowerSystems.get_n_states — Method

get_n_states(value::SimpleFullMachine) -> Int64

Get SimpleFullMachinen_states.

PowerSystems.get_states — Method

get_states(value::SimpleFullMachine) -> Vector{Symbol}

Get SimpleFullMachinestates.

PowerSystems.set_L_1d! — Method

set_L_1d!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_1d.

PowerSystems.set_L_1q! — Method

set_L_1q!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_1q.

PowerSystems.set_L_ad! — Method

set_L_ad!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_ad.

PowerSystems.set_L_aq! — Method

set_L_aq!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_aq.

PowerSystems.set_L_d! — Method

set_L_d!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_d.

PowerSystems.set_L_f1d! — Method

set_L_f1d!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_f1d.

PowerSystems.set_L_ff! — Method

set_L_ff!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_ff.

PowerSystems.set_L_q! — Method

set_L_q!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineL_q.

PowerSystems.set_R! — Method

set_R!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineR.

PowerSystems.set_R_1d! — Method

set_R_1d!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineR_1d.

PowerSystems.set_R_1q! — Method

set_R_1q!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineR_1q.

PowerSystems.set_R_f! — Method

set_R_f!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineR_f.

PowerSystems.set_ext! — Method

set_ext!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineext.

PowerSystems.set_inv_d_fluxlink! — Method

set_inv_d_fluxlink!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineinv_d_fluxlink.

PowerSystems.set_inv_q_fluxlink! — Method

set_inv_q_fluxlink!(value::SimpleFullMachine, val) -> Any

Set SimpleFullMachineinv_q_fluxlink.

SimpleMarconatoMachine

PowerSystems.SimpleMarconatoMachine — Type

mutable struct SimpleMarconatoMachine <: Machine
    R::Float64
    Xd::Float64
    Xq::Float64
    Xd_p::Float64
    Xq_p::Float64
    Xd_pp::Float64
    Xq_pp::Float64
    Td0_p::Float64
    Tq0_p::Float64
    Td0_pp::Float64
    Tq0_pp::Float64
    T_AA::Float64
    ext::Dict{String, Any}
    γd::Float64
    γq::Float64
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 4-states synchronous machine: Simplified Marconato model The derivative of stator fluxes (ψd and ψq) is neglected and ωψd = ψd and ωψq = ψq is assumed (i.e. ω=1.0). This is standard when transmission network dynamics is neglected.

Arguments

R::Float64: Resistance after EMF in machine per unit, validation range: (0, nothing)

Xd::Float64: Reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq::Float64: Reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_p::Float64: Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_p::Float64: Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Xd_pp::Float64: Sub-Transient reactance after EMF in d-axis per unit, validation range: (0, nothing)

Xq_pp::Float64: Sub-Transient reactance after EMF in q-axis per unit, validation range: (0, nothing)

Td0_p::Float64: Time constant of transient d-axis voltage, validation range: (0, nothing)

Tq0_p::Float64: Time constant of transient q-axis voltage, validation range: (0, nothing)

Td0_pp::Float64: Time constant of sub-transient d-axis voltage, validation range: (0, nothing)

Tq0_pp::Float64: Time constant of sub-transient q-axis voltage, validation range: (0, nothing)

T_AA::Float64: Time constant of d-axis additional leakage, validation range: (0, nothing)

ext::Dict{String, Any}

γd::Float64

γq::Float64

states::Vector{Symbol}: The states are:

eq_p: q-axis transient voltage,
ed_p: d-axis transient voltage,
eq_pp: q-axis subtransient voltage,
ed_pp: d-axis subtransient voltage

n_states::Int: SimpleMarconatoMachine has 4 states

internal::InfrastructureSystemsInternal: power system internal reference, do not modify

PowerSystems.get_R — Method

get_R(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineR.

PowerSystems.get_T_AA — Method

get_T_AA(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineT_AA.

PowerSystems.get_Td0_p — Method

get_Td0_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineTd0_p.

PowerSystems.get_Td0_pp — Method

get_Td0_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineTd0_pp.

PowerSystems.get_Tq0_p — Method

get_Tq0_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineTq0_p.

PowerSystems.get_Tq0_pp — Method

get_Tq0_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineTq0_pp.

PowerSystems.get_Xd — Method

get_Xd(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineXd.

PowerSystems.get_Xd_p — Method

get_Xd_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineXd_p.

PowerSystems.get_Xd_pp — Method

get_Xd_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineXd_pp.

PowerSystems.get_Xq — Method

get_Xq(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineXq.

PowerSystems.get_Xq_p — Method

get_Xq_p(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineXq_p.

PowerSystems.get_Xq_pp — Method

get_Xq_pp(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineXq_pp.

PowerSystems.get_ext — Method

get_ext(value::SimpleMarconatoMachine) -> Dict{String, Any}

Get SimpleMarconatoMachineext.

PowerSystems.get_n_states — Method

get_n_states(value::SimpleMarconatoMachine) -> Int64

Get SimpleMarconatoMachinen_states.

PowerSystems.get_states — Method

get_states(value::SimpleMarconatoMachine) -> Vector{Symbol}

Get SimpleMarconatoMachinestates.

PowerSystems.get_γd — Method

get_γd(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineγd.

PowerSystems.get_γq — Method

get_γq(value::SimpleMarconatoMachine) -> Float64

Get SimpleMarconatoMachineγq.

PowerSystems.set_R! — Method

set_R!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineR.

PowerSystems.set_T_AA! — Method

set_T_AA!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineT_AA.

PowerSystems.set_Td0_p! — Method

set_Td0_p!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineTd0_p.

PowerSystems.set_Td0_pp! — Method

set_Td0_pp!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineTd0_pp.

PowerSystems.set_Tq0_p! — Method

set_Tq0_p!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineTq0_p.

PowerSystems.set_Tq0_pp! — Method

set_Tq0_pp!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineTq0_pp.

PowerSystems.set_Xd! — Method

set_Xd!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineXd.

PowerSystems.set_Xd_p! — Method

set_Xd_p!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineXd_p.

PowerSystems.set_Xd_pp! — Method

set_Xd_pp!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineXd_pp.

PowerSystems.set_Xq! — Method

set_Xq!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineXq.

PowerSystems.set_Xq_p! — Method

set_Xq_p!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineXq_p.

PowerSystems.set_Xq_pp! — Method

set_Xq_pp!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineXq_pp.

PowerSystems.set_ext! — Method

set_ext!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineext.

PowerSystems.set_γd! — Method

set_γd!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineγd.

PowerSystems.set_γq! — Method

set_γq!(value::SimpleMarconatoMachine, val) -> Any

Set SimpleMarconatoMachineγq.