Shaft

FiveMassShaft

PowerSystems.FiveMassShaft — Type

mutable struct FiveMassShaft <: Shaft
    H::Float64
    H_hp::Float64
    H_ip::Float64
    H_lp::Float64
    H_ex::Float64
    D::Float64
    D_hp::Float64
    D_ip::Float64
    D_lp::Float64
    D_ex::Float64
    D_12::Float64
    D_23::Float64
    D_34::Float64
    D_45::Float64
    K_hp::Float64
    K_ip::Float64
    K_lp::Float64
    K_ex::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of 5 mass-spring shaft model. It contains a High-Pressure (HP) steam turbine, Intermediate-Pressure (IP) steam turbine, Low-Pressure (LP) steam turbine, the Rotor and an Exciter (EX) mover

Arguments

H::Float64: Rotor inertia constant in MWs/MVA, validation range: (0, nothing)

H_hp::Float64: High pressure turbine inertia constant in MWs/MVA, validation range: (0, nothing)

H_ip::Float64: Intermediate pressure turbine inertia constant in MWs/MVA, validation range: (0, nothing)

H_lp::Float64: Low pressure turbine inertia constant in MWs/MVA, validation range: (0, nothing)

H_ex::Float64: Exciter inertia constant in MWs/MVA, validation range: (0, nothing)

D::Float64: Rotor natural damping in pu, validation range: (0, nothing)

D_hp::Float64: High pressure turbine natural damping in pu, validation range: (0, nothing)

D_ip::Float64: Intermediate pressure turbine natural damping in pu, validation range: (0, nothing)

D_lp::Float64: Low pressure turbine natural damping in pu, validation range: (0, nothing)

D_ex::Float64: Exciter natural damping in pu, validation range: (0, nothing)

D_12::Float64: High-Intermediate pressure turbine damping, validation range: (0, nothing)

D_23::Float64: Intermediate-Low pressure turbine damping, validation range: (0, nothing)

D_34::Float64: Low pressure turbine-Rotor damping, validation range: (0, nothing)

D_45::Float64: Rotor-Exciter damping, validation range: (0, nothing)

K_hp::Float64: High pressure turbine angle coefficient, validation range: (0, nothing)

K_ip::Float64: Intermediate pressure turbine angle coefficient, validation range: (0, nothing)

K_lp::Float64: Low pressure turbine angle coefficient, validation range: (0, nothing)

K_ex::Float64: Exciter angle coefficient, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

states::Vector{Symbol}: (Do not modify.) The states are:

δ: rotor angle,
ω: rotor speed,
δ_hp: rotor angle of high pressure turbine,
ω_hp: rotor speed of high pressure turbine,
δ_ip: rotor angle of intermediate pressure turbine,
ω_ip: rotor speed of intermediate pressure turbine,
δ_lp: rotor angle of low pressure turbine,
ω_lp: rotor speed of low pressure turbine,
δ_ex: rotor angle of exciter,
ω_lp: rotor speed of exciter

n_states::Int: (Do not modify.) FiveMassShaft has 10 states

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

PowerSystems.get_D — Method

get_D(value::FiveMassShaft) -> Float64

Get FiveMassShaftD.

PowerSystems.get_D_12 — Method

get_D_12(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_12.

PowerSystems.get_D_23 — Method

get_D_23(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_23.

PowerSystems.get_D_34 — Method

get_D_34(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_34.

PowerSystems.get_D_45 — Method

get_D_45(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_45.

PowerSystems.get_D_ex — Method

get_D_ex(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_ex.

PowerSystems.get_D_hp — Method

get_D_hp(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_hp.

PowerSystems.get_D_ip — Method

get_D_ip(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_ip.

PowerSystems.get_D_lp — Method

get_D_lp(value::FiveMassShaft) -> Float64

Get FiveMassShaftD_lp.

PowerSystems.get_H — Method

get_H(value::FiveMassShaft) -> Float64

Get FiveMassShaftH.

PowerSystems.get_H_ex — Method

get_H_ex(value::FiveMassShaft) -> Float64

Get FiveMassShaftH_ex.

PowerSystems.get_H_hp — Method

get_H_hp(value::FiveMassShaft) -> Float64

Get FiveMassShaftH_hp.

PowerSystems.get_H_ip — Method

get_H_ip(value::FiveMassShaft) -> Float64

Get FiveMassShaftH_ip.

PowerSystems.get_H_lp — Method

get_H_lp(value::FiveMassShaft) -> Float64

Get FiveMassShaftH_lp.

PowerSystems.get_K_ex — Method

get_K_ex(value::FiveMassShaft) -> Float64

Get FiveMassShaftK_ex.

PowerSystems.get_K_hp — Method

get_K_hp(value::FiveMassShaft) -> Float64

Get FiveMassShaftK_hp.

PowerSystems.get_K_ip — Method

get_K_ip(value::FiveMassShaft) -> Float64

Get FiveMassShaftK_ip.

PowerSystems.get_K_lp — Method

get_K_lp(value::FiveMassShaft) -> Float64

Get FiveMassShaftK_lp.

PowerSystems.get_ext — Method

get_ext(value::FiveMassShaft) -> Dict{String, Any}

Get FiveMassShaftext.

PowerSystems.get_n_states — Method

get_n_states(value::FiveMassShaft) -> Int64

Get FiveMassShaftn_states.

PowerSystems.get_states — Method

get_states(value::FiveMassShaft) -> Vector{Symbol}

Get FiveMassShaftstates.

PowerSystems.set_D! — Method

set_D!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD.

PowerSystems.set_D_12! — Method

set_D_12!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_12.

PowerSystems.set_D_23! — Method

set_D_23!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_23.

PowerSystems.set_D_34! — Method

set_D_34!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_34.

PowerSystems.set_D_45! — Method

set_D_45!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_45.

PowerSystems.set_D_ex! — Method

set_D_ex!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_ex.

PowerSystems.set_D_hp! — Method

set_D_hp!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_hp.

PowerSystems.set_D_ip! — Method

set_D_ip!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_ip.

PowerSystems.set_D_lp! — Method

set_D_lp!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftD_lp.

PowerSystems.set_H! — Method

set_H!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftH.

PowerSystems.set_H_ex! — Method

set_H_ex!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftH_ex.

PowerSystems.set_H_hp! — Method

set_H_hp!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftH_hp.

PowerSystems.set_H_ip! — Method

set_H_ip!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftH_ip.

PowerSystems.set_H_lp! — Method

set_H_lp!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftH_lp.

PowerSystems.set_K_ex! — Method

set_K_ex!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftK_ex.

PowerSystems.set_K_hp! — Method

set_K_hp!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftK_hp.

PowerSystems.set_K_ip! — Method

set_K_ip!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftK_ip.

PowerSystems.set_K_lp! — Method

set_K_lp!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftK_lp.

PowerSystems.set_ext! — Method

set_ext!(value::FiveMassShaft, val) -> Any

Set FiveMassShaftext.

SingleMass

PowerSystems.SingleMass — Type

mutable struct SingleMass <: Shaft
    H::Float64
    D::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
    internal::InfrastructureSystemsInternal
end

Parameters of single mass shaft model. Typically represents the rotor mass

Arguments

H::Float64: Rotor inertia constant in MWs/MVA, validation range: (0, nothing)

D::Float64: Rotor natural damping in pu, validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation.

states::Vector{Symbol}: (Do not modify.) The states are:

δ: rotor angle,
ω: rotor speed

n_states::Int: (Do not modify.) SingleMass has 1 state

internal::InfrastructureSystemsInternal: (Do not modify.) PowerSystems.jl internal reference

PowerSystems.get_D — Method

get_D(value::SingleMass) -> Float64

Get SingleMassD.

PowerSystems.get_H — Method

get_H(value::SingleMass) -> Float64

Get SingleMassH.

PowerSystems.get_ext — Method

get_ext(value::SingleMass) -> Dict{String, Any}

Get SingleMassext.

PowerSystems.get_n_states — Method

get_n_states(value::SingleMass) -> Int64

Get SingleMassn_states.

PowerSystems.get_states — Method

get_states(value::SingleMass) -> Vector{Symbol}

Get SingleMassstates.

PowerSystems.set_D! — Method

set_D!(value::SingleMass, val) -> Any

Set SingleMassD.

PowerSystems.set_H! — Method

set_H!(value::SingleMass, val) -> Any

Set SingleMassH.

PowerSystems.set_ext! — Method

set_ext!(value::SingleMass, val) -> Any

Set SingleMassext.