PowerSimulationsDynamics

API documentation

Index

Exported

PowerSimulationsDynamics.Simulation — Type

Simulation(simulation_folder::String, system::PowerSystems.System, tspan::Tuple{Float64,Float64}) -> Simulation
Simulation(simulation_folder::String, system::PowerSystems.System, tspan::Tuple{Float64,Float64}, perturbations::Array{#s51,1} where #s51<:PowerSimulationsDynamics.Perturbation; kwargs...) -> Simulation

Initializes the simulations and builds the indexing. The input system is not modified during the initialization

Accepted Key Words

system_to_file::Bool: Serializes the original input system

PowerSimulationsDynamics.Simulation! — Function

Simulation!(simulation_folder::String, system::PowerSystems.System, tspan::Tuple{Float64,Float64}) -> Simulation
Simulation!(simulation_folder::String, system::PowerSystems.System, tspan::Tuple{Float64,Float64}, perturbations::Array{#s51,1} where #s51<:PowerSimulationsDynamics.Perturbation; kwargs...) -> Simulation

Initializes the simulations and builds the indexing. The initial conditions are stored in the system.

Accepted Key Words

system_to_file::Bool: Serializes the initialized system

PowerSimulationsDynamics.get_initial_conditions — Method

get_initial_conditions(sim::Simulation) -> Dict{String,Any}

Returns a Dictionary with the resulting initial conditions of the simulation

PowerSimulationsDynamics.get_voltagemag_series — Method

get_voltagemag_series(sim::Simulation, bus_number::Int64) -> Union{Nothing, Tuple{Any,Any}}

Function to obtain the voltage magnitude series out of the DAE Solution. It receives the solution, the dynamical system and the bus number.

PowerSimulationsDynamics.print_device_states — Method

print_device_states(sim::Simulation)

Function to print initial states. It receives the vector of initial states and the dynamical system.

PowerSimulationsDynamics.print_init_states — Method

print_init_states(sim::Simulation)

Function to print initial states. It receives the vector of initial states and the dynamical system.

Internal

PowerSimulationsDynamics.configure_logging — Method

configure_logging(; console_level, file_level, filename) -> InfrastructureSystems.MultiLogger

configure_logging(;
    console_level = Logging.Error,
    file_level = Logging.Info,
    filename = "power-simulations.log",
)

Creates console and file loggers.

Note: Log messages may not be written to the file until flush() or close() is called on the returned logger.

Arguments

console_level = Logging.Error: level for console messages

file_level = Logging.Info: level for file messages

filename::String = power-simulations.log: log file

Example

logger = configure_logging(console_level = Logging.Info)
@info "log message"
close(logger)

PowerSimulationsDynamics.initialize_mach_shaft! — Method

initialize_mach_shaft!(device_states::Any, device::PowerSystems.DynamicGenerator{PowerSystems.AndersonFouadMachine,S<:PowerSystems.Shaft,A<:PowerSystems.AVR,TG<:PowerSystems.TurbineGov,P<:PowerSystems.PSS}) -> Any

Initialitation of model of 6-state (Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

initialize_mach_shaft!(device_states::Any, device::PowerSystems.DynamicGenerator{PowerSystems.BaseMachine,S<:PowerSystems.Shaft,A<:PowerSystems.AVR,TG<:PowerSystems.TurbineGov,P<:PowerSystems.PSS}) -> Any

Initialitation of model of 0-state synchronous (classic model) machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

initialize_mach_shaft!(device_states::Any, device::PowerSystems.DynamicGenerator{PowerSystems.MarconatoMachine,S<:PowerSystems.Shaft,A<:PowerSystems.AVR,TG<:PowerSystems.TurbineGov,P<:PowerSystems.PSS}) -> Any

Initialitation of model of 6-state (Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

initialize_mach_shaft!(device_states::Any, device::PowerSystems.DynamicGenerator{PowerSystems.OneDOneQMachine,S<:PowerSystems.Shaft,A<:PowerSystems.AVR,TG<:PowerSystems.TurbineGov,P<:PowerSystems.PSS}) -> Any

Initialitation of model of 2-state (One d- and One q-) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

initialize_mach_shaft!(device_states::Any, device::PowerSystems.DynamicGenerator{PowerSystems.SimpleAFMachine,S<:PowerSystems.Shaft,A<:PowerSystems.AVR,TG<:PowerSystems.TurbineGov,P<:PowerSystems.PSS}) -> Any

Initialitation of model of 4-state (Simple Anderson-Fouad) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.initialize_mach_shaft! — Method

initialize_mach_shaft!(device_states::Any, device::PowerSystems.DynamicGenerator{PowerSystems.SimpleMarconatoMachine,S<:PowerSystems.Shaft,A<:PowerSystems.AVR,TG<:PowerSystems.TurbineGov,P<:PowerSystems.PSS}) -> Any

Initialitation of model of 4-state (Simple Marconato) synchronous machine in Julia. Refer to Power System Modelling and Scripting by F. Milano for the equations

PowerSimulationsDynamics.kirchoff_laws! — Method

kirchoff_laws!(sys::Any, V_r::Any, V_i::Any, I_injections_r::Any, I_injections_i::Any, dx::Any)

Kirchoff law for buses.
I_gen_r[j]: Real current injection from all generators connected at bus j.
            It is zero if no generator is connected at bus j.
I_gen_i[j]: Real current injection from all generators connected at bus j.
            It is zero if no generator is connected at bus j.