PowerSimulationsDynamics

PowerSimulationsDynamics.BranchImpedanceChange — Type

Use to model changes in the network as sudden impedance changes. Mostly used for small system analysis.

PowerSimulationsDynamics.BranchTrip — Type

Use to model the trip of an AC Branch in in the system. Accepts any ACBranch

PowerSimulationsDynamics.ControlReferenceChange — Type

Use to model control reference changes in devices of the model

PowerSimulationsDynamics.MassMatrixModel — Method

Instantiate a MassMatrixModel for ODE inputs.

PowerSimulationsDynamics.NetworkSwitch — Type

Use to model a change in the network by switching the underlying Ybus in the simulation

PowerSimulationsDynamics.ResidualModel — Method

Instantiate an ResidualModel for ODE inputs.

PowerSimulationsDynamics.ResidualModel — Method

Instantiate an ResidualModel for ForwardDiff calculations

PowerSimulationsDynamics.Simulation — Method

Initializes the simulations and builds the indexing. The input system is not modified during the initialization

Accepted Key Words

initialize_simulation::Bool : Runs the initialization routine. If false, simulation runs based on the operation point stored in System

initial_conditions::Vector{Float64} : Initial conditions provided by the user. Must be used with initialize_simulation = false to not be override.

system_to_file::Bool: Serializes the original input system

console_level::Logging: Sets the level of logging output to the console. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debugg

file_level::Logging: Sets the level of logging output to file. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

PowerSimulationsDynamics.Simulation! — Method

Builds the simulation object and conducts the indexing process. The initial conditions are stored in the system.

Accepted Key Words

initialize_simulation::Bool : Runs the initialization routine. If false, simulation runs based on the operation point stored in System

`initialconditions::Vector{Float64} : Allows the user to pass a vector with the initial condition values desired in the simulation. If initializesimulation = true, these values are used as a first guess and overwritten.

system_to_file::Bool: Serializes the initialized system

console_level::Logging: Sets the level of logging output to the console. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

file_level::Logging: Sets the level of logging output to file. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

PowerSimulationsDynamics.get_activepower_series — Method

Function to obtain the active power output time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.get_imaginary_current_series — Method

Function to obtain the imaginary current time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.get_reactivepower_series — Method

Function to obtain the reactive power output time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.get_real_current_series — Method

Function to obtain the real current time series of a Dynamic Injection series out of the DAE Solution. It receives the solution and the string name of the Dynamic Injection device.

PowerSimulationsDynamics.get_state_series — Method

Function to obtain series of states out of DAE Solution. It receives the solution, the dynamical system and a tuple containing the string name of the Dynamic Injection device and the symbol of the state.

PowerSimulationsDynamics.get_voltage_angle_series — Method

Function to obtain the voltage angle series out of the DAE Solution. It receives the solution, the dynamical system and the bus number.

PowerSimulationsDynamics.get_voltage_magnitude_series — Method

Function to obtain the voltage magnitude series out of the DAE Solution. It receives the solution, the dynamical system and the bus number.

PowerSimulationsDynamics.read_initial_conditions — Method

Returns a Dictionary with the resulting initial conditions of the simulation

PowerSimulationsDynamics.show_states_initial_value — Method

Function to print initial states. It receives the vector of initial states and the dynamical system.

PowerSimulationsDynamics.show_states_initial_value — Method

Function to print initial states. It receives the vector of initial states and the dynamical system.