PowerSimulationsDynamics

PowerSimulationsDynamics.BranchTrip — Type

Use to model the trip of an AC Branch in in the system. Accepts lines or transformer lines

PowerSimulationsDynamics.ControlReferenceChange — Type

Use to model control reference changes in devices of the model

PowerSimulationsDynamics.NetworkSwitch — Type

Use to model a change in the network by switching the underlying Ybus in the simulation

PowerSimulationsDynamics.Simulation — Type

Initializes the simulations and builds the indexing. The input system is not modified during the initialization

Accepted Key Words

initialize_simulation::Bool : Runs the initialization routine. If false, simulation runs based on the operation point stored in System

system_to_file::Bool: Serializes the original input system

console_level::Logging: Sets the level of logging output to the console. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debugg

file_level::Logging: Sets the level of logging output to file. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

PowerSimulationsDynamics.Simulation! — Function

Builds the simulation object and conducts the indexing process. The initial conditions are stored in the system.

Accepted Key Words

initialize_simulation::Bool : Runs the initialization routine. If false, simulation runs based on the operation point stored in System

system_to_file::Bool: Serializes the initialized system

console_level::Logging: Sets the level of logging output to the console. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

file_level::Logging: Sets the level of logging output to file. Can be set to Logging.Error, Logging.Warn, Logging.Info or Logging.Debug

PowerSimulationsDynamics.get_initial_conditions — Method

Returns a Dictionary with the resulting initial conditions of the simulation

PowerSimulationsDynamics.get_state_series — Method

Function to obtain series of states out of DAE Solution. It receives the solution, the dynamical system and a tuple containing the symbol name of the Dynamic Injection device and the symbol of the state.

PowerSimulationsDynamics.get_voltagemag_series — Method

Function to obtain the voltage magnitude series out of the DAE Solution. It receives the solution, the dynamical system and the bus number.

PowerSimulationsDynamics.print_device_states — Method

Function to print initial states. It receives the vector of initial states and the dynamical system.