OuterControl

The outer control is composed by the ReactivePowerControl and the ActivePowerControl types.

PowerSystems.OuterControl — Type

mutable struct OuterControl{
    A <: ActivePowerControl,
    R <: ReactivePowerControl
} <: DynamicInverterComponent
    active_power_control::A
    reactive_power_control::R
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a Outer-Loop controller using a active power controller and a reactive power droop controller.

Arguments

A <: ActivePowerControl: Active power controller (typically droop or virtual inertia).

R <: ReactivePowerControl: Reactive power controller (typically droop).

ext::Dict{String, Any}

states::Vector{Symbol}: Vector of states (will depend on the components).

n_states::Int: Number of states (will depend on the components).

Virtual Inertia

PowerSystems.VirtualInertia — Type

mutable struct VirtualInertia <: ActivePowerControl
    Ta::Float64
    kd::Float64
    kω::Float64
    P_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a Virtual Inertia with SRF using VSM for active power controller

Arguments

Ta::Float64: VSM inertia constant, validation range: (0, nothing)

kd::Float64: VSM damping constant, validation range: (0, nothing)

kω::Float64: frequency droop gain, validation range: (0, nothing)

P_ref::Float64: (default: 1.0) Reference Power Set-point (pu), validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude. See Adding additional fields

states::Vector{Symbol}: (Do not modify.) The states of the VirtualInertia model are:

θ_oc: Phase angle displacement of the virtual synchronous generator model
ω_oc: Speed of the rotating reference frame of the virtual synchronous generator model

n_states::Int: (Do not modify.) VirtualInertia has two states

PowerSystems.get_P_ref — Method

get_P_ref(value::VirtualInertia) -> Float64

Get VirtualInertiaP_ref.

PowerSystems.get_Ta — Method

get_Ta(value::VirtualInertia) -> Float64

Get VirtualInertiaTa.

PowerSystems.get_ext — Method

get_ext(value::VirtualInertia) -> Dict{String, Any}

Get VirtualInertiaext.

PowerSystems.get_kd — Method

get_kd(value::VirtualInertia) -> Float64

Get VirtualInertiakd.

PowerSystems.get_kω — Method

get_kω(value::VirtualInertia) -> Float64

Get VirtualInertiakω.

PowerSystems.get_n_states — Method

get_n_states(value::VirtualInertia) -> Int64

Get VirtualInertian_states.

PowerSystems.get_states — Method

get_states(value::VirtualInertia) -> Vector{Symbol}

Get VirtualInertiastates.

PowerSystems.set_P_ref! — Method

set_P_ref!(value::VirtualInertia, val) -> Any

Set VirtualInertiaP_ref.

PowerSystems.set_Ta! — Method

set_Ta!(value::VirtualInertia, val) -> Any

Set VirtualInertiaTa.

PowerSystems.set_ext! — Method

set_ext!(value::VirtualInertia, val) -> Any

Set VirtualInertiaext.

PowerSystems.set_kd! — Method

set_kd!(value::VirtualInertia, val) -> Any

Set VirtualInertiakd.

PowerSystems.set_kω! — Method

set_kω!(value::VirtualInertia, val) -> Any

Set VirtualInertiakω.

Reactive Power Droop

PowerSystems.ReactivePowerDroop — Type

mutable struct ReactivePowerDroop <: ReactivePowerControl
    kq::Float64
    ωf::Float64
    V_ref::Float64
    ext::Dict{String, Any}
    states::Vector{Symbol}
    n_states::Int
end

Parameters of a Reactive Power droop controller

Arguments

kq::Float64: frequency droop gain, validation range: (0, nothing)

ωf::Float64: filter frequency cutoff, validation range: (0, nothing)

V_ref::Float64: (default: 1.0) Reference Voltage Set-point (pu), validation range: (0, nothing)

ext::Dict{String, Any}: (default: Dict{String, Any}()) An extra dictionary for users to add metadata that are not used in simulation, such as latitude and longitude. See Adding additional fields

states::Vector{Symbol}: (Do not modify.) The states of the ReactivePowerDroop model are:

q_oc: Filtered reactive output power

n_states::Int: (Do not modify.) ReactivePowerDroop has 1 state

PowerSystems.get_V_ref — Method

get_V_ref(value::ReactivePowerDroop) -> Float64

Get ReactivePowerDroopV_ref.

PowerSystems.get_ext — Method

get_ext(value::ReactivePowerDroop) -> Dict{String, Any}

Get ReactivePowerDroopext.

PowerSystems.get_kq — Method

get_kq(value::ReactivePowerDroop) -> Float64

Get ReactivePowerDroopkq.

PowerSystems.get_n_states — Method

get_n_states(value::ReactivePowerDroop) -> Int64

Get ReactivePowerDroopn_states.

PowerSystems.get_states — Method

get_states(value::ReactivePowerDroop) -> Vector{Symbol}

Get ReactivePowerDroopstates.

PowerSystems.get_ωf — Method

get_ωf(value::ReactivePowerDroop) -> Float64

Get ReactivePowerDroopωf.

PowerSystems.set_V_ref! — Method

set_V_ref!(value::ReactivePowerDroop, val) -> Any

Set ReactivePowerDroopV_ref.

PowerSystems.set_ext! — Method

set_ext!(value::ReactivePowerDroop, val) -> Any

Set ReactivePowerDroopext.

PowerSystems.set_kq! — Method

set_kq!(value::ReactivePowerDroop, val) -> Any

Set ReactivePowerDroopkq.

PowerSystems.set_ωf! — Method

set_ωf!(value::ReactivePowerDroop, val) -> Any

Set ReactivePowerDroopωf.