PowerSystems

PowerSystems

API documentation

PowerSystems

Index

Exported

Internal

Index

PowerSystems.PhaseShiftingTransformer

Exported

PowerSystems.PhaseShiftingTransformer — Type.

The 2-W transformer model uses an equivalent circuit assuming the impedance is on the High Voltage Side of the transformer. The model allocates the iron losses and magnetezing suceptance to the primary side

source

Internal